### 2019-10-05 #1787 @whedon
[PRE REVIEW]: Geometric nested sampling: sampling from distributions defined on non-trivial geometries
**Submitting author:** @SuperKam91 (<a href="http://orcid.org/0000-0002-2089-3964">Kamran Javid</a>)
**Repository:** <a href="https://github.com/SuperKam91/gns" target ="_blank">https://github.com/SuperKam91/gns</a>
**Version:** v1.01
