### 2019-10-05 #1787 @kyleniemeyer
Hi @SuperKam91, thanks for your interest in JOSS. Before we start our process, I noticed that while your repository contains a number of Python files, it is not structured like a formal, installable Python package.

Can you work on structuring it in this way, for example following guidelines like https://docs.python-guide.org/writing/structure/ ?

Regarding the paper, I think you need to fix the header:
```yaml
---
title: 'Geometric nested sampling'
tags:
  - Nested sampling
  - Monte Carlo methods
  - Bayesian inference
  - Astrophysics
authors:
  - name: Kamran Javid
    orcid: 0000-0002-2089-3964
    affiliation: "1, 2"
affiliations:
 - name: Astrophysics Group, Cavendish Laboratory, JJ Thomson Avenue, Cambridge CB3 0HE, UK
   index: 1
 - name: Kavli Institute for Cosmology Cambridge, Madingley Road, Cambridge, CB3 0HA, UK
   index: 2
date: 2 October 2019
bibliography: paper.bib
---

# Abstract
```
(I think the missing `---` at the end is the issue)
