+++ paper.md
@@ -1,0 +1,74 @@
+---
+title: 'Geometric nested sampling'
+
+tags:
+  - Nested sampling
+  - Monte Carlo methods
+  - Bayesian inference
+  - Astrophysics
+authors:
+  - name: Kamran Javid
+    orcid: 0000-0002-2089-3964
+    affiliation: "1, 2"
+affiliations:
+ - name: Astrophysics Group, Cavendish Laboratory, JJ Thomson Avenue, Cambridge CB3 0HE, UK
+   index: 1
+ - name: Kavli Institute for Cosmology Cambridge, Madingley Road, Cambridge, CB3 0HA, UK
+   index: 2
+date: 2 October 2019
+bibliography: paper.bib
+---
+
+# Abstract
+
+Metropolis Hastings nested sampling evolves a Markov chain from a current livepoint and accepts new points along the chain according to a version of the Metropolis Hastings acceptance ratio modified to satisfy the likelihood constraint, characteristic of nested sampling algorithms.
+The geometric nested sampling algorithm I present here is a based on the Metropolis Hastings method, but treats parameters as though they represent points on certain geometric objects, namely circles, tori and spheres. For parameters which represent points on a circle or torus, the trial distribution is "wrapped" around the domain of the posterior distribution such that samples cannot be rejected automatically when evaluating the Metropolis ratio due to being outside the sampling domain. Furthermore, this enhances the mobility of the sampler. 
+For parameters which represent coordinates on the surface of a sphere, the algorithm transforms the parameters into a Cartesian coordinate system before sampling which again makes sure no samples are automatically rejected, and provides a physically intutive way of the sampling the parameter space.
+
+# Bayesian inference and nested sampling
+
+For a model $\mathcal{M}$ and data $\vec{\mathcal{D}}$, we can obtain model parameters (also known as input or sampling parameters) $\vec{\theta}$ conditioned on $\mathcal{M}$ and $\vec{\mathcal{D}}$ using Bayes' theorem:
+
+$$
+\mathrm{Pr}\left(\vec{\theta}|\vec{\mathcal{D}},\mathcal{M}\right) = \frac{\mathrm{Pr}\left(\vec{\mathcal{D}}|\vec{\theta},\mathcal{M}\right)\mathrm{Pr}\left(\vec{\theta}|\mathcal{M}\right)}{\mathrm{Pr}\left(\vec{\mathcal{D}}|\mathcal{M}\right)},
+$$
+
+where $\mathrm{Pr}\left(\vec{\theta}|\vec{\mathcal{D}},\mathcal{M}\right) \equiv \mathcal{P}\left(\vec{\theta}\right)$ is the posterior distribution of the model parameter set, $\mathrm{Pr}\left(\vec{\mathcal{D}}|\vec{\theta},\mathcal{M}\right) \equiv \mathcal{L}\left(\vec{\theta}\right)$ is the likelihood function for the data, $\mathrm{Pr}\left(\vec{\theta}|\mathcal{M}\right) \equiv \pi\left(\vec{\theta}\right)$ is the prior probability distribution for the model parameter set, and $\mathrm{Pr}\left(\vec{\mathcal{D}}|\mathcal{M}\right) \equiv \mathcal{Z}$ is the Bayesian evidence of the data given a model $\mathcal{M}$. The evidence can be interpreted as the factor required to normalise the posterior over the model parameter space:
+
+$$
+\mathcal{Z} = \int \mathcal{L}\left(\vec{\theta}\right) \pi\left(\vec{\theta}\right)\, \mathrm{d}\vec{\theta}.
+$$.
+
[…]
