+++ paper.bib
@@ -1,0 +1,140 @@
+@article{chib1995understanding,
+  title={Understanding the metropolis-hastings algorithm},
+  author={Chib, Siddhartha and Greenberg, Edward},
+  journal={The american statistician},
+  volume={49},
+  number={4},
+  pages={327--335},
+  year={1995},
+  publisher={Taylor \& Francis Group}
+}
+
+@article{javid2018physical,
+  title={Physical modelling of galaxy clusters detected by the Planck satellite},
+  author={Javid, Kamran and Olamaie, Malak and Perrott, Yvette C and Carvalho, Pedro and Grainge, Keith JB and Hobson, Michael P and Rumsey, Clare and Saunders, Richard DE},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={483},
+  number={3},
+  pages={3529--3544},
+  year={2018},
+  publisher={Oxford University Press}
+}
+
+@article{javid2018comparison,
+  title={Comparison of physical and observational galaxy cluster modelling},
+  author={Javid, Kamran and Perrott, Yvette C and Hobson, Michael P and Olamaie, Malak and Rumsey, Clare and Saunders, Richard DE},
+  journal={arXiv preprint arXiv:1805.01968},
+  year={2018}
+}
+
+@article{perrott2019sunyaev,
+  title={Sunyaev--Zel’dovich profile fitting with joint AMI-Planck analysis},
+  author={Perrott, Yvette C and Javid, Kamran and Carvalho, Pedro and Elwood, Patrick J and Hobson, Michael P and Lasenby, Anthony N and Olamaie, Malak and Saunders, Richard DE},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={486},
+  number={2},
+  pages={2116--2128},
+  year={2019},
+  publisher={Oxford University Press}
+}
+
+@article{javid2019nested,
+  title={Nested sampling on non-trivial geometries},
+  author={Javid, Kamran},
+  journal={arXiv preprint arXiv:1905.09110},
+  year={2019}
+}
+
+@phdthesis{javid2019physical,
+  title={Physical modelling of galaxy clusters and Bayesian inference in astrophysics},
+  author={Javid, Kamran},
+  year={2019},
+  school={University of Cambridge}
+}
+
+@article{javid2019physical,
+  title={Physical modelling of galaxy cluster Sunyaev--Zel’dovich data using Einasto dark matter profiles},
+  author={Javid, Kamran and Perrott, Yvette C and Rumsey, Clare and Saunders, Richard DE},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={489},
+  number={3},
+  pages={3135--3148},
+  year={2019},
+  publisher={Oxford University Press}
+}
+
+@article{feroz2009bayesian,
+  title={Bayesian modelling of clusters of galaxies from multifrequency-pointed Sunyaev--Zel'dovich observations},
+  author={Feroz, Farhan and Hobson, Michael P and Zwart, Jonathan TL and Saunders, Richard DE and Grainge, Keith JB},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={398},
+  number={4},
+  pages={2049--2060},
+  year={2009},
+  publisher={Blackwell Publishing Ltd Oxford, UK}
+}
+
+@article{feroz2009multinest,
+  title={MultiNest: an efficient and robust Bayesian inference tool for cosmology and particle physics},
+  author={Feroz, F and Hobson, MP and Bridges, M},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={398},
+  number={4},
+  pages={1601--1614},
+  year={2009},
+  publisher={Blackwell Publishing Ltd Oxford, UK}
+}
[…]
