+++ gns/main.py
@@ -1,0 +1,13 @@
+#import standard modules 
+
+#import custom modules
+import tests
+import plotting
+import input
+###########main function
+
+def main():
+	tests.runTests()
+
+if __name__ == '__main__': 
+	main()

[…]
