### 2019-10-08 e5e7071
Updated installation methods
