+++ gns/main.py
@@ -1,13 +1,13 @@
 #import standard modules 
 
 #import custom modules
-import tests
+import nested_run
 import plotting
 import input
 ###########main function
 
 def main():
-	tests.runTests()
+	#call functions in nested_run module here
 
 if __name__ == '__main__': 
 	main()
