+++ requirements.txt
@@ -1,3 +1,2 @@
 numpy
-itertools
 scipy
