### 2019-10-08 #1787 @SuperKam91
Hi @kyleniemeyer thank you for your comments. I have now updated the `paper.md` with the correct formatting. I have also introduced a much more comprehensive installation procedure (and guidance in the `README.md`). The changes can be found at the same url: https://github.com/SuperKam91/gns but note I had to create a new repo due to some issues (the old repo is now at https://github.com/SuperKam91/gns_old). I hope this doesn't cause any issues for  @whedon...!
