+++ paper.md
@@ -27,25 +27,25 @@ For parameters which represent coordinates on the surface of a sphere, the algor
 
 # Bayesian inference and nested sampling
 
-For a model $\mathcal{M}$ and data $\vec{\mathcal{D}}$, we can obtain model parameters (also known as input or sampling parameters) $\vec{\theta}$ conditioned on $\mathcal{M}$ and $\vec{\mathcal{D}}$ using Bayes' theorem:
+For a model $\mathcal{M}$ and data $\mathbf{\mathcal{D}}$, we can obtain model parameters (also known as input or sampling parameters) $\mathbf{\theta}$ conditioned on $\mathcal{M}$ and $\mathbf{\mathcal{D}}$ using Bayes' theorem:
 
 $$
-\mathrm{Pr}\left(\vec{\theta}|\vec{\mathcal{D}},\mathcal{M}\right) = \frac{\mathrm{Pr}\left(\vec{\mathcal{D}}|\vec{\theta},\mathcal{M}\right)\mathrm{Pr}\left(\vec{\theta}|\mathcal{M}\right)}{\mathrm{Pr}\left(\vec{\mathcal{D}}|\mathcal{M}\right)},
+\mathrm{Pr}\left(\mathbf{\theta}|\mathbf{\mathcal{D}},\mathcal{M}\right) = \frac{\mathrm{Pr}\left(\mathbf{\mathcal{D}}|\mathbf{\theta},\mathcal{M}\right)\mathrm{Pr}\left(\mathbf{\theta}|\mathcal{M}\right)}{\mathrm{Pr}\left(\mathbf{\mathcal{D}}|\mathcal{M}\right)},
 $$
 
-where $\mathrm{Pr}\left(\vec{\theta}|\vec{\mathcal{D}},\mathcal{M}\right) \equiv \mathcal{P}\left(\vec{\theta}\right)$ is the posterior distribution of the model parameter set, $\mathrm{Pr}\left(\vec{\mathcal{D}}|\vec{\theta},\mathcal{M}\right) \equiv \mathcal{L}\left(\vec{\theta}\right)$ is the likelihood function for the data, $\mathrm{Pr}\left(\vec{\theta}|\mathcal{M}\right) \equiv \pi\left(\vec{\theta}\right)$ is the prior probability distribution for the model parameter set, and $\mathrm{Pr}\left(\vec{\mathcal{D}}|\mathcal{M}\right) \equiv \mathcal{Z}$ is the Bayesian evidence of the data given a model $\mathcal{M}$. The evidence can be interpreted as the factor required to normalise the posterior over the model parameter space:
+where $\mathrm{Pr}\left(\mathbf{\theta}|\mathbf{\mathcal{D}},\mathcal{M}\right) \equiv \mathcal{P}\left(\mathbf{\theta}\right)$ is the posterior distribution of the model parameter set, $\mathrm{Pr}\left(\mathbf{\mathcal{D}}|\mathbf{\theta},\mathcal{M}\right) \equiv \mathcal{L}\left(\mathbf{\theta}\right)$ is the likelihood function for the data, $\mathrm{Pr}\left(\mathbf{\theta}|\mathcal{M}\right) \equiv \pi\left(\mathbf{\theta}\right)$ is the prior probability distribution for the model parameter set, and $\mathrm{Pr}\left(\mathbf{\mathcal{D}}|\mathcal{M}\right) \equiv \mathcal{Z}$ is the Bayesian evidence of the data given a model $\mathcal{M}$. The evidence can be interpreted as the factor required to normalise the posterior over the model parameter space:
 
 $$
[…]
