### 2019-10-12 #1787 @labarba
@SuperKam91 👋  — I'm the Associate Editor-in-Chief on rotation this week, and have reached out to the potential handling editor by email now. We'll be looking to get started with the review soon. In the meantime, I have two requests:
1. Could you look through our  [list of potential reviewers](https://bit.ly/joss-reviewers) to see if you can suggest someone? ( feel free to suggest reviewers from outside the list, too).
2. Could you think of a modified title for your paper that doesn't sound like a general method, and is more descriptive of your software? It may be as easy as adding "GNS:" as prefix, but if you think of a few more descriptive words, that is useful.

Thanks!
