### 2019-10-14 SuperKam91/gns#1 @ziotom78
Unable to install the package
The README says that the following command should be enough to install the package, yet it fails:

```
$ pip install GNS==1.0
ERROR: Could not find a version that satisfies the requirement GNS==1.0 (from versions: none)
ERROR: No matching distribution found for GNS==1.0
```

(I was trying to install the package to perform the [review](https://github.com/openjournals/joss-reviews/issues/1809) for the JOSS.)
