### 2019-10-14 SuperKam91/gns#1 @SuperKam91
> The README says that the following command should be enough to install the package, yet it fails:
>
> ```
> $ pip install GNS==1.0
> ERROR: Could not find a version that satisfies the requirement GNS==1.0 (from versions: none)
> ERROR: No matching distribution found for GNS==1.0
> ```
>
> (I was trying to install the package to perform the [review](https://github.com/openjournals/joss-reviews/issues/1809) for the JOSS.)

Hi @ziotom78 thank you for your message. That is very odd, I just tried `pip installing` the package on a machine which has never had the package, and it seemed to work fine.
![Screenshot from 2019-10-15 00-00-59](https://user-images.githubusercontent.com/25101227/66788313-104b5180-eedf-11e9-9931-5727ac44d62d.png)

Perhaps try `pip install GNS` instead? Lemme know how you get on.
