### 2019-10-14 SuperKam91/gns#1 @ziotom78
Mmm… I think I spotted the problem. I am running Python 3.7; does this package require Python 2.7?
