### 2019-10-15 SuperKam91/gns#1 @SuperKam91
> Mmm… I think I spotted the problem. I am running Python 3.7; does this package require Python 2.7?

HI @ziotom78 yes thanks for spotting this. I just realised the `wheel` file I built and put on PyPi is a `py2`. I developed and tested the code using `python 2.7` so I recommend using that version when using the algorithm. I have edited the `README` and `setup.py` accordingly.
