### 2019-10-15 SuperKam91/gns#1 @ziotom78
Are you really sure you want to publish a library that requires a version of Python that in a few weeks will be [deprecated and no longer maintained](https://pythonclock.org/)?
