### 2019-10-15 SuperKam91/gns#1 @ziotom78
I was able to build a Singularity container to run Python 2.7, as my system (Linux Manjaro) does no longer supports it. However, `pip` fails:

```
$ singularity exec python2.sif sh
Singularity> pip install GNS==1.0
Collecting GNS==1.0
  Downloading https://files.pythonhosted.org/packages/62/55/1cec82e159e361e765563e6094795269e0406f48ea347a9ac17f000270de/GNS-1.0-py2-none-any.whl (83kB)
    100% |################################| 92kB 2.6MB/s
Collecting scipy (from GNS==1.0)
  Downloading https://files.pythonhosted.org/packages/ee/5b/5afcd1c46f97b3c2ac3489dbc95d6ca28eacf8e3634e51f495da68d97f0f/scipy-1.3.1.tar.gz (23.6MB)
    100% |################################| 23.6MB 79kB/s
    Complete output from command python setup.py egg_info:
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-build-blvcH3/scipy/setup.py", line 31, in <module>
        raise RuntimeError("Python version >= 3.5 required.")
    RuntimeError: Python version >= 3.5 required.

Command "python setup.py egg_info" failed with error code 1 in /tmp/pip-build-blvcH3/scipy/
You are using pip version 8.1.1, however version 19.3 is available.
You should consider upgrading via the 'pip install --upgrade pip' command.
```

The problem lies in the fact you did not specify version numbers for the dependencies in `requirements.txt`, and the latest version of `scipy` no longer works with Python2. Therefore, when `pip` tries to install the dependencies, it fails. On your system nothing happens because you probably have an old version of `numpy` and `scipy`.

I reiterate my suggestion: you should really consider converting this package to Python3. It is not very difficult, most of the process can be automatized using `2to3`. It will be available to a much wider audience.
