### 2019-10-15 f710013
Made python 3 compatible
