+++ text_output/th_Sc_mhwg.txt
@@ -1,2500 +1,0 @@
-3.044596168249014248e-06 -8.262711373422909844e+00 3.149913990652743134e+00
-3.039468230929081671e-06 -8.262399065781776741e+00 3.156606057778450936e+00
-3.033663801451345892e-06 -8.262312574745783067e+00 3.125202119619189034e+00
-3.034214944782280347e-06 -8.260132912773324065e+00 3.178452230247246657e+00
-3.041720783142538847e-06 -8.255664231514137086e+00 3.081643580082702005e+00
-3.036356680292044082e-06 -8.255431294951856458e+00 3.080679349466620476e+00
-3.031815060062982749e-06 -8.254930158650712002e+00 3.204530392900010849e+00
-3.032403078889306579e-06 -8.252738225356029744e+00 3.212711918417907597e+00
-3.029821966482825767e-06 -8.251591762365055871e+00 3.066548051650228146e+00
-3.039054585481786148e-06 -8.246551145128666604e+00 3.051288175537367753e+00
-3.055308949947931041e-06 -8.239218901108838367e+00 3.250345274564279130e+00
-3.049798934182246501e-06 -8.239025950008342036e+00 3.250788804312422897e+00
-3.050730200624971546e-06 -8.236722640566737752e+00 3.255951111939115350e+00
-3.050281927052942970e-06 -8.234871588463564152e+00 3.259937383986578396e+00
-3.065308960291943037e-06 -8.227959240234387295e+00 3.273769548604501090e+00
-3.065290589719291234e-06 -8.225967230647107442e+00 3.005689284584640575e+00
-3.061461962672427815e-06 -8.225219034472452506e+00 3.278869707796375543e+00
-3.070565883086787307e-06 -8.220251727908271633e+00 3.287664481552285967e+00
-3.136185385192843953e-06 -8.197108380679766881e+00 3.323144788833375607e+00
-3.166452227002509878e-06 -8.185505804077559233e+00 2.944622344384896007e+00
-3.198074120350074124e-06 -8.173570800033377992e+00 2.929916996590300915e+00
-3.225554560299406664e-06 -8.163016696462097599e+00 3.365480158009835243e+00
-3.228989088681816350e-06 -8.159954473339128711e+00 3.368902556761089961e+00
-3.227072033808878421e-06 -8.158550348241401906e+00 2.912730295138520109e+00
-3.223477611804384335e-06 -8.157666800183061895e+00 3.371426615376336589e+00
-3.253945226903833955e-06 -8.146261401846476957e+00 2.899560348474226945e+00
-3.280026235187101873e-06 -8.136280154398882303e+00 2.889359732235829537e+00
-3.299680348221115207e-06 -8.128307973027549593e+00 2.881493674398727389e+00
-3.303063982605004483e-06 -8.125285052691086918e+00 2.878571148088787890e+00
-3.316459743960790855e-06 -8.119239695506987431e+00 2.872819684210865532e+00
-3.316764528901544834e-06 -8.117149796376343929e+00 2.870859131167212475e+00
-3.326273758234307650e-06 -8.112288876049561281e+00 3.416833555482692297e+00
-3.320626414782895525e-06 -8.111990115037361448e+00 2.866077065248104905e+00
-3.314219718386033701e-06 -8.111923339794133270e+00 2.866015707329342188e+00
-3.317003935274144063e-06 -8.109085607634421322e+00 3.419764923089807951e+00
-3.316883429090662896e-06 -8.107123935460005981e+00 2.861639988065312945e+00
-3.317823116301897597e-06 -8.104842668658642779e+00 3.423602009535348323e+00
-3.333608076531189832e-06 -8.098098322117488479e+00 2.853586499322992687e+00
[…]
