### 2019-10-15 2d2ab38
Made python 3 compatible
