### 2019-10-15 c33ce4d
Made python 3 compatible
