+++ docs/index.rst
@@ -1,0 +1,187 @@
+.. Geometric nested sampling documentation master file, created by
+   sphinx-quickstart on Thu Oct 17 22:36:17 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+=====================================================
+Welcome to Geometric nested sampling's documentation!
+=====================================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+Version
+=======
+
+Version: 1.0
+
+Overview
+========
+
+Nested sampling algorithms are a family of algorithms designed to sample from posterior probability distributions associated with e.g. Bayesian inference. The overall idea in nested sampling is to convert an n-dimensional integral over the parameter space into a single dimensional integral over 'prior volume' and equivalently sample from this domain. The purpose of the geometric nested sampling algorithm presented here is to efficiently sample the distributions of parameters which exhibit certain geometrical properties e.g. circular, toroidal or spherical. Such parameters are rife in the astrophysics community, where many time- or angle-periodic parameters are present.
+
+The algorithm
+=============
+
+Can run Markov Chain Monte Carlo (MCMC) Metropolis Hastings (MH) nested sampling algorithm in 'vanilla' (see https://projecteuclid.org/euclid.ba/1340370944 and https://arxiv.org/abs/0704.3704) and/or 'geometric' (https://arxiv.org/abs/1905.09110) sampling modes. See the appropriate papers for detailed explanations of how these algorithms operate.
+
+In vanilla mode, the trial points are sampled in the physical (native) space of the input parameters.
+In geometric mode (main point of the algorithm), parameter values represent points on a geometric object (e.g. circle, torus or sphere), and trial points are picked by sampling from spaces parameterised by Cartesian coordinates. If the sampled points do not lay on the original object (in general they will not), then they are projected back onto the object before being transformed back to the native space for evaluation of the MH acceptance ratio.
+
+The sampling mode for each parameter is specified in ``paramGeomList`` passed through ``setupDict`` to ``nestedRun()`` (``"vanilla"``, ``"wrapped"`` or ``"sphere"``). ``"wrapped"`` parameters use a wrapped trial distribution. ``"sphere"`` parameters are transformed into their Euclidean representation and sampled in this space. Geometric sampling works best for parameters with 'periodic' sampling domains e.g. angles, as sampling in geometric space results in wrapping parameter values at end of domains. The sampler also works well for parameters which exhibit 'spherical' properties. Note for sampling spherical sampling, two physical parameters are required for each geometric transformation. See ``tests.py`` for examples of how to specify what parameters are sampled in what way and the below section showing ``paramGeomList`` examples. 
+
+Main output of geometric nested sampling run is file of sampled points and their weights. These are stored in ``text_output/file_name.txt``, whose format follows that explained in https://getdist.readthedocs.io/en/latest/intro.html. Similarly, metadata files are also generated for each run e.g. ``file_name.paramnames``, see https://getdist.readthedocs.io/en/latest/intro.html for more details on these files. The format of the text outputs is such that they can be plotted very easily using ``getdist``, a kernel density esimate plotter.
+
+Toy models
+==========
+
[…]
