+++ PACKAGEAPI.md
@@ -1,5 +1,0 @@
-# Sampler
-
-# Plotting
-
-# Misc
