### 2019-10-18 d334939
Added more documentation
