+++ gns/theoretical_funcs.py
@@ -14,6 +14,14 @@ def nDIntegratorZTheor(integrandFuncs, limitsList, integrandLogVal, divByFact = 
 	If integrandLogVal evaluates to true, does integration method which takes exp of log of integrand * some large number given by integrandLogVal
 	to avoid underflow. The final integral result (and error) is then divided by exp(integrandLogVal) to get the final value.
 	If divByFact == True, divides integral by exp(integrandLogVal) to give integral of function. If not leaves this step out, and essentially gives integral of function * exp(integrandLogVal). The latter can be useful when exp(integrandLogVal) would under/overflow, in which case one should consider log(integral of function) which is given by log(integral of function * exp(integrandLogVal)) - integrandLogVal
+	
+	Args:
+
+	integrandFuncs : list functions which form integrand (through their product)
+	limitsList : list integration limits
+	integrandLogVal : boolean whether to take log of integrand or not 
+	divByFact : boolean whether to divide by log of integrandLogVal or not
+
 	"""
 	if integrandLogVal:
 		if divByFact: 
@@ -28,6 +36,15 @@ def nDIntegratorHTheor(integrandFuncs, limitsList, integrandLogVal, divByFact = 
 	"""
 	As above but has to call nquad with a  slightly different function representing integrand when using
 	exp(log(integrand)) method for calculating, due to LLhood(theta) part of integrand
+
+	Args:
+
+	integrandFuncs : list functions which form integrand (through their product)
+	limitsList : list integration limits
+	integrandLogVal : boolean whether to take log of integrand or not 
+	divByFact : boolean whether to divide by log of integrandLogVal or not
+	LLhoodFunc : function or False log likelihood function if to be included in integrand calculation
+
 	"""
 	if integrandLogVal:
 		if divByFact:
@@ -105,7 +122,13 @@ def integrateLogFunc(logPriorFunc, LLhoodFunc, targetSupport):
 	Uses logaddexp to get log of integrand.
 	Takes equally spaced samples in each dimension (so width in each dimension is upper - lower lim / num samples in dim).
 	priorHyperParams is as explained in getTheoreticalSamples()
-	TODO: consider multidimensional trapezium rule. Probably won't bother
+	
+	Args:
+
+	logPriorFunc : function log prior function
+	LLhoodFunc : function or False log likelihood function if to be included in integrand calculation
+	targetSupport : array target support values in array of shape (3, nDims)
+
 	"""
 	sampleWidth = 1.
 	oneDn = 100 #number of points per dimension
@@ -141,6 +164,15 @@ def getPriorIntegrandAndLimits(priorFunc, targetSupport, integrandFuncs, integra
 	Not much slower than integrateAll = False, as that method still has to integrate Lhood over all dimensions (just not prior)
 	integrateAll = False assumes all finitely bounded prior dimensions are rectangular, and works out their prior volume
 	as being a rectangle. Also uses priorFuncsPdf, so can't be used with user defined functions. NOTE THIS WILL NOT GIVE THE CORRECT RESULT IF ANY OF THE TOY PRIORS ARE BOUNDED BUT AREN'T UNIFORM, AS WHEN CHECKING TARGETSUPPORT IT WILL ASSUME THAT DIMENSION IS UNIFORM. Could be fixed by adding a flag to targetSupport as to whether prior is uniform or not, but this method is sort of deprecated anyway.
+	
+	Args:
+
+	priorFunc : function prior function
+	targetSupport : array target support values in array of shape (3, nDims)
+	integrandFuncs : list functions which form integrand (through their product)
[…]
