### 2019-10-18 f16a21c
Update conf.py
