+++ docs/requirements.txt
@@ -2,3 +2,4 @@
 cython>=0.28.3
 numpy>=1.14.5
 numpydoc
+sphinx_rtd_theme
