+++ docs/nested_run.rst
@@ -1,0 +1,3 @@
+.. automodule:: gns.nested_run
+   :members:
+   :undoc-members:
