+++ docs/conf.py
@@ -137,21 +137,21 @@ pygments_style = "sphinx"
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-html_theme_options = {
-    #    'canonical_url': '',
-    #    'analytics_id': '',
-    "logo_only": False,
-    "display_version": True,
-    "prev_next_buttons_location": "top",
-    #    'style_external_links': False,
-    #    'vcs_pageview_mode': '',
-    # Toc options
-    "collapse_navigation": False,
-    "sticky_navigation": True,
-    "navigation_depth": 4,
-    "includehidden": True,
-    "titles_only": False,
-}
+# html_theme_options = {
+#     #    'canonical_url': '',
+#     #    'analytics_id': '',
+#     "logo_only": False,
+#     "display_version": True,
+#     "prev_next_buttons_location": "top",
+#     #    'style_external_links': False,
+#     #    'vcs_pageview_mode': '',
+#     # Toc options
+#     "collapse_navigation": False,
+#     "sticky_navigation": True,
+#     "navigation_depth": 4,
+#     "includehidden": True,
+#     "titles_only": False,
+# }
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
