+++ docs/conf.py
@@ -22,7 +22,6 @@
 # is needed in order to build the documentation
 import os
 import sys
-# import sphinx_rtd_theme
 import numpydoc
 
 # local module should not be added to sys path if it's installed on RTFD
@@ -59,10 +58,10 @@ extensions = [
     "sphinx.ext.ifconfig",
     "sphinx.ext.viewcode",
     "sphinx.ext.autosummary",
-#     'sphinx.ext.githubpages',
+    'sphinx.ext.githubpages',
     "sphinx.ext.napoleon",  # parameters look better than with numpydoc only
     "numpydoc",
-#     "sphinx_rtd_theme"
+    "sphinx_rtd_theme"
 ]
 
 # autosummaries from source-files
@@ -131,27 +130,27 @@ pygments_style = "sphinx"
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-# html_theme = "sphinx_rtd_theme"
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-# html_theme_options = {
-#     #    'canonical_url': '',
-#     #    'analytics_id': '',
-#     "logo_only": False,
-#     "display_version": True,
-#     "prev_next_buttons_location": "top",
-#     #    'style_external_links': False,
-#     #    'vcs_pageview_mode': '',
-#     # Toc options
-#     "collapse_navigation": False,
-#     "sticky_navigation": True,
-#     "navigation_depth": 4,
-#     "includehidden": True,
-#     "titles_only": False,
-# }
+html_theme_options = {
+    #    'canonical_url': '',
+    #    'analytics_id': '',
+    "logo_only": False,
+    "display_version": True,
+    "prev_next_buttons_location": "top",
+    #    'style_external_links': False,
+    #    'vcs_pageview_mode': '',
+    # Toc options
+    "collapse_navigation": False,
+    "sticky_navigation": True,
+    "navigation_depth": 4,
+    "includehidden": True,
+    "titles_only": False,
+}
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
