+++ docs/nested_run.rst
@@ -1,3 +1,2 @@
-.. automodule:: gns.nested_run
-   :members:
-   :undoc-members:
+.. currentmodule::
+   gns
