+++ docs/nested_run.rst
@@ -1,2 +1,1 @@
-.. currentmodule::
-   gns
+
