+++ docs/conf.py
@@ -29,16 +29,22 @@ import numpydoc
 # sys.path.insert(0, os.path.abspath("../../"))
 # from pentapy import __version__ as ver
 
+#Location of Sphinx files
+sys.path.insert(0, os.path.abspath('./../'))
 
-# def skip(app, what, name, obj, skip, options):
-#     if name in ["__call__"]:
-#         return False
-#     return skip
-
-
-# def setup(app):
-#     app.connect("autodoc-skip-member", skip)
-
+import sphinx.apidoc
+def setup(app):
+    app.add_javascript('copybutton.js')
+    sphinx.apidoc.main(['-f', #Overwrite existing files
+                        '-T', #Create table of contents
+                        '-e', #Give modules their own pages
+                        #'-E', #user docstring headers
+                        #'-M', #Modules first
+                        '-o', #Output the files to:
+                        './_autogen/', #Output Directory
+                        './../gns', #Main Module directory
+                        ]
+    )
 
 # -- General configuration ------------------------------------------------
 
