+++ docs/requirements.txt
@@ -3,3 +3,6 @@ cython>=0.28.3
 numpy>=1.14.5
 numpydoc
 sphinx_rtd_theme
+scipy
+matplotlib
+GNS==1.0
