### 2019-10-19 25acca9
Modified docstrings
