+++ docs/index.rst
@@ -179,19 +179,3 @@ Alternatively, one can run ``python setup.py install`` or ``pip install .`` from
 Note that by default, the tests in the ``tests/`` directory run the gns algorithm from the code in the repo in which ``tests/`` sits. To change this i.e. to use the installed package, in ``tests.py``, set ``run_repo = False``.
 
 
-.. currentmodule::
-   gns
-
-.. toctree::
-   :glob:
-   :maxdepth: 2
-   :caption: Quick start:
-
-   _pages/*
-
-.. toctree::
-   :glob:
-   :maxdepth: 2
-   :caption: Packages:
-
-   _autogen/gns
