+++ docs/index.rst
@@ -132,11 +132,6 @@ Requirements
 
 This package was developed and tested using ``python 3.6`` and thus I recommend using this version when running the algorithm. The two ``wheel`` files on PyPi are ``py2`` or ``py3`` files, the appropriate one will be selected by ``pip`` depending on whether it is called using ``python 2.X`` or ``python 3.X``. Note however, I have not fully tested the ``python 2.X`` version and so I warn against using it.
 
-Package API
-===========
-
-Please see ``PACKAGEAPI.md`` in the root of the repo.
-
 Installation
 ============
 
