+++ docs/gns.rst
@@ -1,0 +1,4 @@
+.. toctree::
+   :maxdepth: 4
+
+   gns
