+++ docs/nested_run.rst
@@ -1,4 +1,1 @@
-.. toctree::
-   :maxdepth: 4
 
-   gns
