+++ docs/index.rst
@@ -8,6 +8,7 @@ Welcome to Geometric nested sampling's documentation!
 =====================================================
 
 .. toctree::
+   :glob:
    :maxdepth: 1
    :caption: Package API:
    _autogen/gns
