### 2019-10-22 SuperKam91/gns#3 @ziotom78
Example in README fails
Hi @SuperKam91, I have spotted another problem while reading the documentation.

Running the example provided in the [documentation](https://geometric-nested-sampling.readthedocs.io/en/latest/index.html), the following errors appears:

```
AttributeError: module 'scipy.misc' has no attribute 'logsumexp'
```

This seems to be due to [line 71 in ns_loop_funcs.py](https://github.com/SuperKam91/gns/blob/master/gns/ns_loop_funcs.py#L71), where `scipy.misc.logsumexp` appears while it should be [`scipy.special.logsumexp`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.special.logsumexp.html).

This bug is related to the review currently running on JOSS: https://github.com/openjournals/joss-reviews/issues/1809
