### 2019-10-22 #1809 @ziotom78
I am still completing the review of the code, but I started reading the paper as well. One thing I note is Fig. 2, which has two "X" marks that are explained neither in the caption nor in the text. The figure is obviously Fig. 3 in the [arXiv paper 1905.09110](https://arxiv.org/pdf/1905.09110.pdf); however, in the arXiv paper Fig. 3 makes sense because it is better explained and placed alongside Fig. 2. I believe that you should include the same figure and the same level of detail in this paper as well.

If the paper length is an issue, you could scale down Fig. 2 in the JOSS paper a bit: the information content is very small (a sphere, three points, and three labels), but the space it takes is really huge.
