### 2019-10-22 f48a475
fixed logsumexp
