+++ gns/keeton_calculations.py
@@ -1,6 +1,10 @@
 #import standard modules 
 import numpy as np
 import scipy
+try: #newer scipy versions
+	import scipy.special.logsumexp as logsumexp
+except ImportError: #older scipy versions
+	import scipy.misc.logsumexp as logsumexp
 
 #import custom modules
 from . import calculations
@@ -100,7 +104,7 @@ def calcEofZKeetonLog(LLhoods, nLive, nest):
 	logEoftArr = np.log(calculations.getEofftArr(calculations.EoftPowi, nLive, nest))
 	logLEoft = LLhoods + logEoftArr
 	#return tools.logAddArr2(-np.inf, logLEoft) - np.log(nLive)
-	return scipy.misc.logsumexp(logLEoft) - np.log(nLive)
+	return logsumexp(logLEoft) - np.log(nLive)
 
 def calcEofZ2KeetonLog(LLhoods, nLive, nest):
 	"""
@@ -120,7 +124,7 @@ def calcLogSums(LLhoods, nLive, nest):
 	innerLogSums = np.fromiter(calcInnerLogSums(LLhoods, nLive, nest), dtype = float, count = nest)
 	outerLogSums = logLEoft + innerLogSums
 	#return tools.logAddArr2(-np.inf, outerLogSums)
-	return scipy.misc.logsumexp(outerLogSums)
+	return logsumexp(outerLogSums)
 
 def calcInnerLogSums(LLhoods, nLive, nest):
 	"""
@@ -130,7 +134,7 @@ def calcInnerLogSums(LLhoods, nLive, nest):
 		logEoft2OverEoftArr = np.log(calculations.getEofftArr(calculations.Eoft2OverEoftPowi, nLive, k))
 		innerLogTerms = LLhoods[:k] + logEoft2OverEoftArr
 		#innerLogSum = tools.logAddArr2(-np.inf, innerLogTerms)
-		innerLogSum = scipy.misc.logsumexp(innerLogTerms)
+		innerLogSum = logsumexp(innerLogTerms)
 		yield innerLogSum
 
 def calcLogSumsLoop(LLhoods, nLive, nest):
@@ -165,7 +169,7 @@ def calcHKeetonLog2(logEofZ, LLhoods, nLive, nest):
 	"""
 	logSumTerms = np.log(LLhoods) + LLhoods + np.log(calculations.getEofftArr(calculations.EoftPowi, nLive, nest))
 	#logSumTerm = np.log(1. / logEofZ) + np.log(1. / nLive) + tools.logAddArr2(-np.inf, logSumTerms)
-	logSumTerm = np.log(1. / logEofZ) + np.log(1. / nLive) + scipy.misc.logsumexp(logSumTerms)
+	logSumTerm = np.log(1. / logEofZ) + np.log(1. / nLive) + logsumexp(logSumTerms)
 	maxTerm = max(logSumTerm, logEofZ)
 	logH = tools.logSubExp(logSumTerm, np.log(logEofZ), logSumTerm)
 	return np.exp(logH)
@@ -241,7 +245,7 @@ def calcEofZFinalKeetonLog(finalLLhoods, nLive, nest):
 	Based on function calcEofZFinalKeeton().
 	"""
 	#logLhoodAv = tools.logAddArr2(-np.inf, finalLLhoods) - np.log(nLive)
-	logLhoodAv = scipy.misc.logsumexp(finalLLhoods) - np.log(nLive)
+	logLhoodAv = logsumexp(finalLLhoods) - np.log(nLive)
[…]
