### 2019-10-22 SuperKam91/gns#3 @SuperKam91
Hi @ziotom78 my apologies for this- over time I have switched between different versions of `scipy`, which have had the `logsumexp` function in different modules (`scipy.special` or `scipy.misc`). I have now fixed the code to work for both.
