### 2019-10-22 bcd4b4d
Update paper.md
