+++ paper.md
@@ -62,8 +62,8 @@ This solution can be problematic if computing the likelihood is computationally 
 Another issue which most sampling algorithms are subject to occurs when the modes of the posterior distribution are far away from each other in $\mathbf{\theta}$ space, e.g. when they are at "opposite ends" of the domain of $\pi$. In the context of nested sampling this can result in one or more of the modes not being sampled accurately, particularly in the case of low livepoint runs.
 Thus a sampling algorithm should be able to efficiently manoeuvre between well separated modes which lie at the "edges" of $\pi$'s support. 
 
-Specific to the geometric nested sampler, is the way it treats parameters that exhibit certain geometrical properties. For example, parameters which exhibit circular or periodic properties (e.g. time or angle), the trial distribution used in the MCMC evaluation is 'wrapped' around its domain as shown in Figure 1. A wrapped trial distribution such as this one means that samples are never automatically rejected for being outside of the domain of the posterior. Furthermore, any modes occurring near the edge of the domain are still "close to the other end" of the domain, and so can be sampled more efficiently.
-For parameters which exhibit spherical properties (e.g. polar and azimuthal angles $\theta$ and $\phi$), the geometric nested sampler uses another trick. Instead of considering their joint rectangular sampling space with area $\pi \times 2\pi$, the algorithm transforms their values into their three-dimensional Euclidean representation ($x, y, z$), and samples from a spherical Gaussian trial distribution. Any sampled points which do not lay on the original sphere described by $\theta$ and $\phi$ are projected back onto the sphere before being converted back to their angular forms, which ensures no trial points outside of the domain are sampled. This idea is illustrated in Figure 2.
+Specific to the geometric nested sampler, is the way it treats parameters that exhibit certain geometrical properties. For example, parameters which exhibit circular or periodic properties (e.g. time or angle), the trial distribution used in the MCMC evaluation is 'wrapped' around its domain as shown in Figure 1. A wrapped trial distribution such as this one ensures that samples are never automatically rejected for being outside of the domain of the posterior. Furthermore, any modes occurring near the edge of the domain are still "close to the other end" of the domain, and so can be sampled more efficiently.
+For parameters which exhibit spherical properties (e.g. polar and azimuthal angles $\theta$ and $\phi$), the geometric nested sampler uses another trick. Instead of considering their joint rectangular sampling space with area $\pi \times 2\pi$, the algorithm transforms their values into their three-dimensional Euclidean representation ($x, y, z$), and samples from a spherical Gaussian trial distribution. Any sampled points which do not lay on the original sphere described by $\theta$ and $\phi$ are projected back onto the sphere before being converted back to their angular forms, which ensures no trial points outside of the domain are sampled. This idea is illustrated in Figure 2: at iteration $l$ of the algorithm we consider a point ($\theta_{l}$, $\phi_{l}$) as a point in the MCMC chain. This point can be transformed into its Cartesian representation ($x_{l}, y_{l}, z_{l}$), which forms the centre of the trial distribution from which a candidate point is sampled. In general the point sampled from the trial distribution, ($x', y', z'$), will not lay on the surface of the sphere. Thus the point is projected back onto the sphere to give ($x_{t}, y_{t}, z_{t}$), which can be converted back into angular coordinates to give the trial point ($\theta_{t}$, $\phi_{t}$).
 
 <img src="paper_figs/Sphere.jpg" alt="Sampling spherical parameters by using their Euclidean interpretation. $(x_{l}, y_{l}, z_{l})$ represents the centre of the trial distribution, from which point $(x', y', z')$ is sampled. Since this point does not lay on the sphere, it is projected back onto it (to give point $(x_{t}, y_{t}, z_{t})$) before it is converted back to its angular representation." width="500"/>
 
