### 2019-10-22 #1809 @zhampel
@SuperKam91 As I begin my review, can you please include some guidelines for contributing or reporting issues to your codebase? Here's an [example](https://jrbourbeau.github.io/pyunfold/contributing.html).
