### 2019-10-22 #1809 @SuperKam91
> I am still completing the review of the code, but I started reading the paper as well. One thing I note is Fig. 2, which has two "X" marks that are explained neither in the caption nor in the text. The figure is obviously Fig. 3 in the [arXiv paper 1905.09110](https://arxiv.org/pdf/1905.09110.pdf); however, in the arXiv paper Fig. 3 makes sense because it is better explained and placed alongside Fig. 2. I believe that you should include the same figure and the same level of detail in this paper as well.
>
> If the paper length is an issue, you could scale down Fig. 2 in the JOSS paper a bit: the information content is very small (a sphere, three points, and three labels), but the space it takes is really huge.

Hi @ziotom78 thank you for your comments on the paper. While I gave some information on the coordinate labels in the caption of Figure 2 (n.b. captions should be visible in the `.pdf`s generated in this thread, but probably not in the `.md` file itself), I have added a couple of sentences in the main text elaborating more on them. With regards to the figure sizes, I have tried to reduce them using different syntaxes to the one suggested in the JOSS `paper.md` example. These worked in the `.md` files, but not in the `.pdf` generated here, and so I have had to revert back to the original sizing. It would be very useful if the JOSS template provided an example of how to scale images which is compatible with the `pdf` generation used here.
