### 2019-10-23 #1809 @SuperKam91
Hi @zhampel , I have added a section explaining how to contribute to the project/ report issues at the bottom of the `README` of the repo.
