### 2019-10-29 SuperKam91/gns#3 @ziotom78
The code is still not working, sorry… You are writing

```python
import scipy.special.logsumexp as logsumexp
```
This causes the error
```
In [7]: import scipy.special.logsumexp
<ipython-input-7-75ec627707c9> in <module>
----> 1 import scipy.special.logsumexp

ModuleNotFoundError: No module named 'scipy.special.logsumexp'
```

The syntax you have used is valid only if `logsumexp` is a (sub)module, not if it is an identifier. You should write

```python
from scipy.special import logsumexp
```
