### 2019-10-29 SuperKam91/gns#4 @ziotom78
Specify the requisites for the example in the README
Ideally, examples in a README should be self-contained and easy to run. The example provided in the README has the following problems:

- It requires the user to run `pip install getdist`, but this is not stated clearly
- It requires the user to create two directories, `../text_output/` and `../image_output/`

None of these actions is too difficult for users with average experience, but it might lead people evaluating your software running away from it. Often people run examples before having completed reading the documentation.

Possible ways to solve the two points above:

- Clearly state that the user should install `getdist` before trying the example, and provide the `pip` command to do so
- Insert a call to `os.mkdir' in the example, so that you are sure the paths are always present. You should check if the path already exists, because people are likely to re-run the same example over and over again, changing the parameters to check how this affects the result.

(This issue is related to the [ongoing review of this package](https://github.com/openjournals/joss-reviews/issues/1809) for JOSS.)
