### 2019-10-29 797848f
Fixed logsumexp imports
