### 2019-10-29 SuperKam91/gns#4 @SuperKam91
@ziotom78 thank you for your suggestions. I have added a few sentences detailing the `getdist` installation at the beginning of the example. I have also added a few lines of code to check for the mentioned directories, and create them if they don't exist
