+++ README.md
@@ -52,8 +52,8 @@ shape = 'circle' #run circle toy model, simple von Mises distribution
 sampler = 'MH WG' #geometric nested sampler
 
 #make sure directories to store output exist, and if not, create them
-textDir = '../text_output/'
-plotDir = '../image_output/'
+textDir = './text_output/'
+plotDir = './image_output/'
 if not os.path.exists(textDir):
   os.mkdir(textDir)
 if not os.path.exists(plotDir):
