### 2019-11-02 SuperKam91/gns#3 @ziotom78
Unfortunately the example still fails. This seems to be related to the line specifying the variable `paramNames`, as everything works if I comment it out:

```
ParseException                            Traceback (most recent call last)
~/tmp/venv/lib/python3.7/site-packages/matplotlib/mathtext.py in parse(self, s, fonts_object, fontsize, dpi)
   2597         try:
-> 2598             result = self._expression.parseString(s)
   2599         except ParseBaseException as err:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in parseString(self, instring, parseAll)
   1938                 # catch and re-raise exception from here, clears out pyparsing internal stack trace
-> 1939                 raise exc
   1940         else:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in parseString(self, instring, parseAll)
   1928         try:
-> 1929             loc, tokens = self._parse(instring, 0)
   1930             if parseAll:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseCache(self, instring, loc, doActions, callPreParse)
   1829                 try:
-> 1830                     value = self._parseNoCache(instring, loc, doActions, callPreParse)
   1831                 except ParseBaseException as pe:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseNoCache(self, instring, loc, doActions, callPreParse)
   1668                 try:
-> 1669                     loc, tokens = self.parseImpl(instring, preloc, doActions)
   1670                 except IndexError:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in parseImpl(self, instring, loc, doActions)
   4429         if self.expr is not None:
-> 4430             return self.expr._parse(instring, loc, doActions, callPreParse=False)
   4431         else:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseCache(self, instring, loc, doActions, callPreParse)
   1829                 try:
-> 1830                     value = self._parseNoCache(instring, loc, doActions, callPreParse)
   1831                 except ParseBaseException as pe:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseNoCache(self, instring, loc, doActions, callPreParse)
   1668                 try:
-> 1669                     loc, tokens = self.parseImpl(instring, preloc, doActions)
   1670                 except IndexError:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in parseImpl(self, instring, loc, doActions)
   4036             else:
-> 4037                 loc, exprtokens = e._parse(instring, loc, doActions)
   4038             if exprtokens or exprtokens.haskeys():

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseCache(self, instring, loc, doActions, callPreParse)
   1829                 try:
-> 1830                     value = self._parseNoCache(instring, loc, doActions, callPreParse)
   1831                 except ParseBaseException as pe:

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in _parseNoCache(self, instring, loc, doActions, callPreParse)
   1672             else:
-> 1673                 loc, tokens = self.parseImpl(instring, preloc, doActions)
   1674

~/tmp/venv/lib/python3.7/site-packages/pyparsing.py in parseImpl(self, instring, loc, doActions)
   3782         if loc < len(instring):
-> 3783             raise ParseException(instring, loc, self.errmsg, self)
   3784         elif loc == len(instring):

ParseException: Expected end of text, found '$'  (at char 0), (line:1, col:1)

<ipython-input-3-74b20f1b8d23> in <module>
     45 #plot data using getdist
     46 #NOTE, calls to output.writeParamNames() and output.writeRanges() are made in nested_run to create .paramnames and .ranges files required by getdist
---> 47 plotting.callGetDist([outputFile], plotFile + '.png', nDims, ['mhwg'])

~/tmp/gns/gns/plotting.py in callGetDist(chainsFilePrefix, plotName, nParams, plotLegend)
    120         g = getdist.plots.getSubplotPlotter(width_inch = 6)
    121 	g.triangle_plot(chains, paramList,
--> 122 	filled = False, legend_labels = plotLegend)
    123         if save:
    124                 g.export(plotName)

~/tmp/venv/lib/python3.7/site-packages/getdist/plots.py in triangle_plot(self, roots, params, legend_labels, plot_3d_with_param, filled, shaded, contour_args, contour_colors, contour_ls, contour_lws, line_args, label_order, legend_ncol, legend_loc, upper_roots, title_limit, upper_kwargs, diag1d_kwargs, markers, marker_args, param_limits, **kwargs)
   2125                          legend_ncol=legend_ncol or (None if upper_roots is None else len(labels)),
   2126                          legend_loc=legend_loc, no_gap=self.settings.no_triangle_axis_labels,
-> 2127                          no_extra_legend_space=upper_roots is None)
   2128         if self.settings.no_triangle_axis_labels: plt.subplots_adjust(wspace=0, hspace=0)
   2129

~/tmp/venv/lib/python3.7/site-packages/getdist/plots.py in finish_plot(self, legend_labels, legend_loc, line_offset, legend_ncol, label_order, no_gap, no_extra_legend_space, no_tight)
   1662                 warnings.simplefilter("ignore")
   1663                 if no_gap:
-> 1664                     plt.tight_layout(h_pad=0, w_pad=0)
   1665                 else:
   1666                     plt.tight_layout()

~/tmp/venv/lib/python3.7/site-packages/matplotlib/pyplot.py in tight_layout(pad, h_pad, w_pad, rect)
   1373         labels) will fit into. Default is (0, 0, 1, 1).
   1374     """
-> 1375     gcf().tight_layout(pad=pad, h_pad=h_pad, w_pad=w_pad, rect=rect)
   1376
   1377

~/tmp/venv/lib/python3.7/site-packages/matplotlib/figure.py in tight_layout(self, renderer, pad, h_pad, w_pad, rect)
   2474         kwargs = get_tight_layout_figure(
   2475             self, self.axes, subplotspec_list, renderer,
-> 2476             pad=pad, h_pad=h_pad, w_pad=w_pad, rect=rect)
   2477         if kwargs:
   2478             self.subplots_adjust(**kwargs)

~/tmp/venv/lib/python3.7/site-packages/matplotlib/tight_layout.py in get_tight_layout_figure(fig, axes_list, subplotspec_list, renderer, pad, h_pad, w_pad, rect)
    360                                      subplot_list=subplot_list,
    361                                      ax_bbox_list=ax_bbox_list,
--> 362                                      pad=pad, h_pad=h_pad, w_pad=w_pad)
    363
    364     # kwargs can be none if tight_layout fails...

~/tmp/venv/lib/python3.7/site-packages/matplotlib/tight_layout.py in auto_adjust_subplotpars(fig, renderer, nrows_ncols, num1num2_list, subplot_list, ax_bbox_list, pad, h_pad, w_pad, rect)
    109             continue
    110
--> 111         tight_bbox_raw = union([ax.get_tightbbox(renderer) for ax in subplots
    112                                 if ax.get_visible()])
    113         tight_bbox = TransformedBbox(tight_bbox_raw,

~/tmp/venv/lib/python3.7/site-packages/matplotlib/tight_layout.py in <listcomp>(.0)
    110
    111         tight_bbox_raw = union([ax.get_tightbbox(renderer) for ax in subplots
--> 112                                 if ax.get_visible()])
    113         tight_bbox = TransformedBbox(tight_bbox_raw,
    114                                      fig.transFigure.inverted())

~/tmp/venv/lib/python3.7/site-packages/matplotlib/axes/_base.py in get_tightbbox(self, renderer, call_axes_locator, bbox_extra_artists)
   4353
   4354         if self.axison:
-> 4355             bb_xaxis = self.xaxis.get_tightbbox(renderer)
   4356             if bb_xaxis:
   4357                 bb.append(bb_xaxis)

~/tmp/venv/lib/python3.7/site-packages/matplotlib/axis.py in get_tightbbox(self, renderer)
   1176               if a.get_visible()),
   1177             *ticklabelBoxes,
-> 1178             *ticklabelBoxes2,
   1179         ]
   1180         bboxes = [b for b in bboxes

~/tmp/venv/lib/python3.7/site-packages/matplotlib/axis.py in <genexpr>(.0)
   1174             *(a.get_window_extent(renderer)
   1175               for a in [self.label, self.offsetText]
-> 1176               if a.get_visible()),
   1177             *ticklabelBoxes,
   1178             *ticklabelBoxes2,

~/tmp/venv/lib/python3.7/site-packages/matplotlib/text.py in get_window_extent(self, renderer, dpi)
    888             raise RuntimeError('Cannot get window extent w/o renderer')
    889
--> 890         bbox, info, descent = self._get_layout(self._renderer)
    891         x, y = self.get_unitless_position()
    892         x, y = self.get_transform().transform_point((x, y))

~/tmp/venv/lib/python3.7/site-packages/matplotlib/text.py in _get_layout(self, renderer)
    296             if clean_line:
    297                 w, h, d = renderer.get_text_width_height_descent(
--> 298                     clean_line, self._fontproperties, ismath=ismath)
    299             else:
    300                 w = h = d = 0

~/tmp/venv/lib/python3.7/site-packages/matplotlib/backends/backend_agg.py in get_text_width_height_descent(self, s, prop, ismath)
    204         if ismath:
    205             ox, oy, width, height, descent, fonts, used_characters = \
--> 206                 self.mathtext_parser.parse(s, self.dpi, prop)
    207             return width, height, descent
    208

~/tmp/venv/lib/python3.7/site-packages/matplotlib/mathtext.py in parse(self, s, dpi, prop)
   3369             self.__class__._parser = Parser()
   3370
-> 3371         box = self._parser.parse(s, font_output, fontsize, dpi)
   3372         font_output.set_canvas_size(box.width, box.height, box.depth)
   3373         return font_output.get_results(box)

~/tmp/venv/lib/python3.7/site-packages/matplotlib/mathtext.py in parse(self, s, fonts_object, fontsize, dpi)
   2601                                         err.line,
   2602                                         " " * (err.column - 1) + "^",
-> 2603                                         str(err)]))
   2604         self._state_stack = None
   2605         self._em_width_cache = {}

ValueError:
$$\phi$$
^
Expected end of text, found '$'  (at char 0), (line:1, col:1)
```
