### 2019-11-06 #1809 @ziotom78
As issue [#3](https://github.com/SuperKam91/gns/issues/3) is now closed, I think that the paper and the package are ok, but let's wait for @zhampel's approval too. The functionality provided by the package looks good, nice work!

However, there are a few suggestions I would like to give to @SuperKam91 :

- The API documentation is ok, but it is surely not «good» [according to JOSS' guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation):

    - It is just a list of functions provided by the package, with no example for any of them. Compare this with the documentation of the Python's standard library (e.g., the [datetime module](https://docs.python.org/3/library/datetime.html)): short code snippet explain how each function is meant to be used.

    - As the list of functions provided by the package is very long, some text introducing each module (e.g., `gns.keeton_calculations`) would surely have helped interested readers. Again, Python's documentation is a good example to be used as reference.

- The README is overly long, and it provides long textual descriptions of how to use the library, when short Python snippets would have provided the same information in a much clearer form.

- The review process was quite long because there were trivial errors in the source code that should have been spotted by the author.

In the future, please read carefully [Author and Reviewer Guides](https://joss.readthedocs.io/en/latest/index.html) before submitting a paper to JOSS, as this will ensure a much faster and smoother review process.
