### 2019-11-08 #1809 @zhampel
@SuperKam91
I agree with @ziotom78 as well. Your submission's functionality is as advertised, but the documentation is lacking in detail. For example, while not explicitly required, function input types and descriptors provide users context for implementation. This is also reflected in the difficulty in scanning the example code in the README, which I would not characterize as a usage guide.

With respect to testing, it doesn't appear that the scripts in the `test` directory strictly qualify as unit tests, specifically the lack of assertion type tests. This would also be facilitated or at least clarified by care detailing of input argument types and type-checks as appropriate. This is also evident in abundant commented code ending with `#debug` or similar. For example on line 120 of `nested_run.py`:

```       #elif setupDict['sampler'] == 'MH' or setupDict['sampler'] == 'MH geom': #temporary whilst testing geometric ns```

If such lines are intended for debugging or testing, then the preference would be to add in some testing capability or flags to ensure such tests are performed with requisite feedback.

Furthermore, I think the paper and docs still have some errors. Examples:
- on page 3: "For example, parameters which... the trial distribution..."
  The sentence is not complete as 'For example' and 'the trial distribution' do not relate properly.
  You can simply state 'periodic properties' instead of 'circular or periodic'
- In the 4th paragraph of the online docs: "Note for sampling spherical sampling, two physical parameters..."
- You also use quotes frequently when they are not necessary.

I recommend that the written components of the submission be re-reviewed and corrected, as well as addressing the lack of clear comments as well as copious commented code within the main codebase. I also highly recommend use of a style guide for Python code such as [pep8](https://www.python.org/dev/peps/pep-0008/).
