### 2019-11-08 6c92f82
Added unit tests for GNS
