+++ gns/nested_run.py
@@ -17,7 +17,7 @@ from . import geom_sampler
 
 ############nested run functions
 
-def NestedRun(priorFunc, invPriorFunc, LhoodFunc, paramNames, targetSupport, setupDict, LLhoodFunc = None):
+def NestedRun(priorFunc, invPriorFunc, LhoodFunc, paramNames, targetSupport, setupDict, LLhoodFunc = None, return_vals = False):
 	"""
 	Wrapper around linear and log nested run functions.
 
@@ -37,16 +37,21 @@ def NestedRun(priorFunc, invPriorFunc, LhoodFunc, paramNames, targetSupport, set
 	
 	LLhoodFunc : function log likelihood function
 
+	return_vals : bool whether to return statistics parameters of nested run or not, 
+	i.e. the expected log evidence E[ln(Z)], its variance var[ln(Z)] and the K-L divergence H.
+	If False, writes these values to file instead (and prints to stdout).
+
 	"""
 	if setupDict['space'] == 'linear':
-		NestedRunLinear(priorFunc, invPriorFunc, LhoodFunc, paramNames, targetSupport, setupDict)
+		return NestedRunLinear(priorFunc, invPriorFunc, LhoodFunc, paramNames, targetSupport, setupDict, return_vals)
 	elif setupDict['space'] == 'log':
-		NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport, setupDict)
+		return NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport, setupDict, return_vals)
 
-def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport, setupDict):
+def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport, setupDict, return_vals = False):
 	"""
 	function which completes a NS run. parameters of priors and likelihood need to be specified, as well as a flag indication type of prior for each dimension and the pdf for the lhood.
-	setupDict contains other setup parameters such as termination type & factor, method of finding new livepoint, details of how weights are calculated, how final Z contribution is added, and directory/file prefix for saved files.
+	setupDict contains other setup parameters such as termination type & factor, method of finding new livepoint, details of how weights are calculated, how final Z contribution is added,
+	and directory/file prefix for saved files.
 
 	Args:
 
@@ -61,6 +66,10 @@ def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport,
 	targetSupport : array target support values in array of shape (3, nDims)
 	
 	setupDict : dict setup dictionary explained in README.md
+
+	return_vals : bool whether to return statistics parameters of nested run or not, 
+	i.e. the expected log evidence E[ln(Z)], its variance var[ln(Z)] and the K-L divergence H.
+	If False, writes these values to file instead (and prints to stdout).
 
 	"""
 	nLive = 500 #low value is 50, high value is 500
@@ -90,14 +99,10 @@ def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames, targetSupport,
 	nest = 0
 	logZLive = np.inf
 	checkTermination = 100
[…]
