+++ tests/tests.py
@@ -3,6 +3,7 @@ import numpy as np
 import itertools
 import scipy
 import matplotlib.pyplot as plt
+import time
 
 run_repo = True
 #import custom modules
@@ -18,40 +19,233 @@ import gns.theoretical_funcs as theoretical_funcs
 import gns.output as output
 import gns.input as input
 
-def runTests():
+TOL = 1e-6
+
+def assert_same(x, y, tol):
 	"""
-	Run geometric nested sampling tests specified in body of the function.
+	Asserts the floats x and y are within tol of each other
+	i.e. |x - y| < tol. Returns boolean according to this,
+	and |x - y| itself
 	"""
-	#configuration for toy model runs
+	diff = np.abs(x - y)
+	return diff < tol, diff
+
+def test_gns_stats_circle():
+	"""
+	Test correct stats are generated for geometric nested sampler for circle toy model. 
+	Note these values have been compared with the theoretically correct values.
+	"""
+	ElnZTrue = -1.778287448059121
+	VarlnZTrue = 0.00222452082689939
+	HTrue = 1.009856152283123
+	paramGeomLists = [['wrapped']] #wrap parameter geometric nested sampler for von Mises distribution (circle)
+	doPlots = [False]
+	return_vals = True
+	shapeList = ['circle'] #run circle toy model, simple von Mises distribution
+	shapeSuffixes = ['_Sc']
+	samplerList = ['MH WG'] #just do geometric MH nested sampling
+	samplerSuffixes = ['_mhwg'] 
+	ElnZPred, VarlnZPred, HPred = runTests(shapeList, shapeSuffixes, samplerList, samplerSuffixes, paramGeomLists, doPlots, return_vals = True)
+	bool_same, diff = assert_same(ElnZTrue, ElnZPred, TOL)
+	if bool_same:
+		print("E[ln(Z)] test for circle model complete.")
+	else:
+		print("E[ln(Z)] test for circle model failed. Discrepancy between true and obtained value is " + str(diff))
+	bool_same, diff = assert_same(VarlnZTrue, VarlnZPred, TOL)
+	if bool_same:
+		print("Var[ln(Z)] test for circle model complete.")
+	else:
+		print("Var[ln(Z)] test for circle model failed. Discrepancy between true and obtained value is " + str(diff))
+	bool_same, diff = assert_same(HTrue, HPred, TOL)
+	if bool_same:
+		print("K-L divergence test for circle model complete.")
+	else:
+		print("K-L divergence test for circle model failed. Discrepancy between true and obtained value is " + str(diff))
+
+def test_gns_stats_torus():
+	"""
+	Test correct stats are generated for geometric nested sampler for torus toy model. 
+	Note these values have been compared with the theoretically correct values.
+	"""
+	ElnZTrue = -3.5870812483323076
+	VarlnZTrue = 0.004308045833100849
+	HTrue = 2.013937631676229
+	paramGeomLists = [['wrapped', 'wrapped']] #wrap parameter geometric nested sampler for torus
+	doPlots = [True]
+	shapeList = ['torus'] #run torus toy model, simple von Mises distribution
+	shapeSuffixes = ['_St']
+	samplerList = ['MH WG'] #just do geometric MH nested sampling
+	samplerSuffixes = ['_mhwg'] 
+	ElnZPred, VarlnZPred, HPred = runTests(shapeList, shapeSuffixes, samplerList, samplerSuffixes, paramGeomLists, doPlots, return_vals = True)
+	bool_same, diff = assert_same(ElnZTrue, ElnZPred, TOL)
+	if bool_same:
+		print("E[ln(Z)] test for torus model complete.")
+	else:
[…]

[…]
