### 2019-11-08 afc3e86
Update paper.md
