+++ paper.md
@@ -49,6 +49,8 @@ The integral extends over the region(s) of the parameter space contained within 
 
 Nested sampling has been used extensively within the astrophysics community for both computing the posterior distributions of parameters of interest and for model comparison (see e.g. [@feroz2009bayesian; @javid2018physical; @javid2018comparison; @perrott2019sunyaev; @javid2019physical] which all use data from the Arcminute MicroKelvin Imager [@hickish2018digital] to solve Bayesian inference problems). Two of the most popular nested sampling algorithms are Multinest [@feroz2009multinest] and PolyChord  [@handley2015polychord; @handley2015polychord]. The former of these approximates the iso-likelihood contours as hyperellipsoids in the unit hypercube in which it samples. PolyChord performs coordinate transformations such that the resultant space is easier to navigate.
 
+A comparison between the performance of the geometric nested sampler and MultiNest is presented in @javid2019nested. Figures 6 - 9 of the paper show their performance on a toroidal model, while Figure 11 shows their relative performance on a complex spherical toy model.
+
 # Summary of the geometric nested sampler
 
 Here I present the geometric nested sampling algorithm, as described in @javid2019nested. The geometric nested sampling algorithm is based on the Metropolis Hastings nested sampling algorithm presented in @feroz2008multimodal, which at each iteration, evolves an MCMC chain to find the next livepoint (point within the iso-likelihood contour). Each point in the chain is sampled from a trial distribution, and whether that point is accepted or not, depends on the value of the Metropolis Hastings ratio [@chib1995understanding]. In the case that the point is not accepted, the next point in the chain is taken to be the point from the previous step.
