### 2019-11-08 cae77ae
Added extra documentation for toy models
