+++ gns/toy_models.py
@@ -8,8 +8,23 @@ from . import geom
 def getToyHypersGeom(shape):
 	"""
 	wrapper around getToyHypers() to get priors/ lhoods
-	specific to geometric nested sampling project
-	NOTE 'sphere' changed to l8 on 18th Feb 2018
+	specific to geometric nested sampling project.
+	NOTE 'sphere' changed to l8 on 18th Feb 2018.
+	See getToyHypers for priors/likelihood that these different shapes (models) correspond to.
+
+	Args:
+
+	shape: string representing toy model name e.g. 'torus III'
+
+	Returns:
+
+	values of respective dictionaries, which contain lists of parameter names, prior and likelihood types, and their hyperparameters in the form:
+	paramNames: list of strings representing parameter names
+	priorParams: list containing prior type (denoted by integer) and hyperparameters (array)
+	LhoodParams: list containing likelihood type (denoted by integer) and hyperparameters (array)
+	Priors which prior types integers correspond to can be found in fitPriors() in prob_funcs.py.
+	Likelihoods which likelihood types integers correspond to can be found in fitLhood() in prob_funcs.py.
+
 	"""
 	shapeDict = {'circle': ['n1', 'p1', 'l4'], 'torus': ['n2', 'p2', 'l5'], 'torus II': ['n4', 'p7', 'l9'], 'torus III': ['n5', 'p8', 'l10'], 'torus IV': ['n11', 'p16', 'l21'], 'torus V': ['n12', 'p17', 'l22'], 'sphere': ['n2', 'p3', 'l8'], 'sphere II': ['n2', 'p3', 'l7'], 'sphere III': ['n2', 'p9', 'l11'], 'sphere IV': ['n2', 'p9', 'l12'], 'sphere V': ['n2', 'p3', 'l11'], 'sphere VI': ['n2', 'p3', 'l12'], '3 sphere IV': ['n6', 'p10', 'l13'], '5 sphere IV': ['n7', 'p11', 'l14'], '6 sphere IV': ['n8', 'p12', 'l15'], '10d gauss sphere IV': ['n9', 'p13', 'l16'], '20d gauss sphere IV': ['n10', 'p14', 'l17'], 'sphere VII': ['n2', 'p9', 'l18'], '20d gauss sphere VII': ['n10', 'p14', 'l19'], '6 sphere VIII': ['n8', 'p15', 'l20']}
 	namesPriorLhood = shapeDict[shape]
@@ -19,6 +34,22 @@ def getToyHypersGen(dists):
 	"""
 	wrapper around getToyHypers() to get priors/ lhoods
 	specific to geometric nested sampling project
+
+	See getToyHypers for priors/likelihood that these different models correspond to.
+
+	Args:
+
+	shape: string representing toy model name e.g. 'gauss p gauss l'
+
+	Returns:
+
+	values of respective dictionaries, which contain lists of parameter names, prior and likelihood types, and their hyperparameters in the form:
+	paramNames: list of strings representing parameter names
+	priorParams: list containing prior type (denoted by integer) and hyperparameters (array)
+	LhoodParams: list containing likelihood type (denoted by integer) and hyperparameters (array)
+	Priors which prior types integers correspond to can be found in fitPriors() in prob_funcs.py.
+	Likelihoods which likelihood types integers correspond to can be found in fitLhood() in prob_funcs.py.
+
 	"""
 	genDict = {'uniform p gauss l': ['n1', 'p1', 'l1'], 'uniform uniform p gauss l': ['n2', 'p2', 'l2'], 'gauss p gauss l': ['n1', 'p6', 'l1']}
 	namesPriorLhood = genDict[dists]
[…]
