### 2019-11-08 #1809 @SuperKam91
> @SuperKam91
> I agree with @ziotom78 as well. Your submission's functionality is as advertised, but the documentation is lacking in detail. For example, while not explicitly required, function input types and descriptors provide users context for implementation.

Hi @zhampel thank you for your comments. I will address them sequentially:

> This is also reflected in the difficulty in scanning the example code in the README, which I would not characterize as a usage guide.

I'm not quite sure I follow unfortunately, are you saying the example needs more comments to explain what each line of code does? Anyway, I have added more documentation to the `toy_models.py` sub-package which is called from the example.

> With respect to testing, it doesn't appear that the scripts in the `test` directory strictly qualify as unit tests, specifically the lack of assertion type tests.

While I didn't see anything in the JOSS guidelines specifying that unit tests were strictly necessary, I appreciate your point, and also acknowledge the fact that a lot of JOSS repos do include them, and therefore I have included some in `tests.py`. For a number of toy models, these essentially check that the statistics associated with the nested sampling runs are calculated correctly.

This would also be facilitated or at least clarified by care detailing of input argument types and type-checks as appropriate. This is also evident in abundant commented code ending with `#debug` or similar. For example on line 120 of `nested_run.py`:
>
> ` #elif setupDict['sampler'] == 'MH' or setupDict['sampler'] == 'MH geom': #temporary whilst testing geometric ns`
>
> If such lines are intended for debugging or testing, then the preference would be to add in some testing capability or flags to ensure such tests are performed with requisite feedback.

I have taken care of such redundant/commented out code.

> Furthermore, I think the paper and docs still have some errors. Examples:
>
>     * on page 3: "For example, parameters which... the trial distribution..."
>       The sentence is not complete as 'For example' and 'the trial distribution' do not relate properly.
>       You can simply state 'periodic properties' instead of 'circular or periodic'
>
>     * In the 4th paragraph of the online docs: "Note for sampling spherical sampling, two physical parameters..."

I have addressed these two typos, thank you for pointing them out.

> I recommend that the written components of the submission be re-reviewed and corrected, as well as addressing the lack of clear comments as well as copious commented code within the main codebase. I also highly recommend use of a style guide for Python code such as [pep8](https://www.python.org/dev/peps/pep-0008/).

I appreciate the code API may not be very good in certain places, but as I said to @ziotom78 in a closed issue, I have ensured that the API is far superior for all the sub-packages/functions which an end-user is typically going to interact with, i.e. the sub-packages:
`nested_run.py`
`plotting.py`
`output.py`
`theoretical_funcs.py`

While I would love to re-write the package following pep8, I think it is unlikely I will get time to.

On another note, I added an extra sentence to the paper explaining where the relative performances of GNS and another popular nested sampling algorithm, MultiNest, can be found (at the end of the section Bayesian inference and nested sampling).
