### 2019-11-08 f7a7a92
Added extra documentation for prob funcs
