+++ gns/nested_sampler.py
@@ -1745,37 +1745,24 @@ def NestedRun(priorParams, LLhoodParams, paramNames, setupDict):
 	EofZ = np.exp(logEofZ)
 	EofZ2 = np.exp(logEofZ2)
 	varZ = calcVariance(EofZ, EofZ2)
-	#logVarZ = calcVarianceLog(logEofZ, logEofZ2)
-	#logEofZK, logEofZ2K = calcZMomentsKeetonLog(np.array(deadPointsLLhood), nLive, nest)
 	EofZK, EofZ2K = calcZMomentsKeeton(np.exp(np.array(deadPointsLLhood)), nLive, nest)
 	varZK = calcVariance(EofZK, EofZ2K)
-	#logVarZK = calcVarianceLog(logEofZK, logEofZ2K)
-	#HK = calcHKeetonLog(logEofZK, np.array(deadPointsLLhood), nLive, nest)
 	HK = calcHKeeton(EofZK, np.exp(np.array(deadPointsLLhood)), nLive, nest)
 	if setupDict['verbose']:
 		printBreak()
 		printZHValues(EofZ, EofZ2, varZ, H, 'linear', 'before final', 'recursive')
 		printZHValues(EofZK, EofZ2K, varZK, HK, 'linear', 'before final', 'Keeton equations')
-		#printZHValues(logEofZ, logEofZ2, logVarZ, H, 'log', 'before final', 'recursive')
-		#printZHValues(logEofZK, logEofZ2K, logVarZK, HK, 'log', 'before final', 'Keeton equations')
 	logEofZTotal, logEofZ2Total, H, livePointsPhysFinal, livePointsLLhoodFinal, logEofXFinalArr = getFinalContributionLog(setupDict['verbose'], setupDict['ZLiveType'], setupDict['trapezoidalFlag'], nFinal, logEofZ, logEofZ2, logEofX, logEofWeights, H, livePointsPhys, livePointsLLhood, avLLhood, liveLLhoodMax, liveMaxIndex, LLhoodStar)
 	totalPointsPhys, totalPointsLLhood, logEofXArr, logEofWeights = getTotal(deadPointsPhys, livePointsPhysFinal, deadPointsLLhood, livePointsLLhoodFinal, logEofXArr, logEofXFinalArr, logEofWeights)
 	EofZTotal = np.exp(logEofZTotal)
 	EofZ2Total = np.exp(logEofZ2Total)
 	varZ = calcVariance(EofZTotal, EofZ2Total)
-	#logEofZFinalK, logEofZ2FinalK = calcZMomentsFinalKeetonLog(livePointsLLhood, nLive, nest)
 	EofZFinalK, EofZ2FinalK = calcZMomentsFinalKeeton(np.exp(livePointsLLhood), nLive, nest)
 	varZFinalK = calcVariance(EofZFinalK, EofZ2FinalK)
-	#logVarZFinalK = calcVarianceLog(logEofZFinalK, logEofZ2FinalK)
-	#logEofZZFinalK = calcEofZZFinalKeetonLog(np.array(deadPointsLLhood), livePointsLLhood, nLive, nest)
 	EofZZFinalK = calcEofZZFinalKeeton(np.exp(np.array(deadPointsLLhood)), np.exp(livePointsLLhood), nLive, nest)
-	#logVarZTotalK = getVarTotalKeetonLog(logVarZK, logVarZFinalK, logEofZK, logEofZFinalK, logEofZZFinalK)
 	varZTotalK = getVarTotalKeeton(varZK, varZFinalK, EofZK, EofZFinalK, EofZZFinalK)
-	#logEofZTotalK = getEofZTotalKeetonLog(logEofZK, logEofZFinalK)
[…]
