### 2019-11-08 6cb10d7
Added extra documentation for prob funcs
