### 2019-11-08 aca2566
Added extra documentation for prob funcs
