### 2019-11-08 #1809 @zhampel
@SuperKam91 I pulled your most recent changes, but I would not agree with you statement that
> I have taken care of such redundant/commented out code.

Looking again in `nested_run.py`, I still see a significant number of commented full lines with qualifications like 'temporary' and 'debug'.

It would be worth you intended users' time to clarify your code as much as possible.
