### 2019-11-09 #1809 @ziotom78
> While I would love to re-write the package following pep8, I think it is unlikely I will get time to.

No need to rewrite your package, there are automatic tools like [autopep8](https://github.com/hhatto/autopep8), [yapf](https://github.com/google/yapf) and [black](https://github.com/psf/black) that can do the job for you.
