+++ gns/geom_sampler.py
@@ -1,474 +1,603 @@
-#import standard modules 
+# import standard modules
 import numpy as np
 
-#import custom modules
+# import custom modules
 from . import geom
 
+
 def splitGeomParams(paramGeomList):
-	"""
-	Get indexes of parameters from paramList
-	based on whether parameter should be sampled in 
-	physical (non-geometric) or geometric space,
-	and add these to two lists which can be used
-	to index parameter array 
-	"""
-	nonGeomList = []
-	geomList = []
-	shapeList = []
-	boundaryList = []
-	for i, geo in enumerate(paramGeomList):
-		if 'vanilla' in geo or 'wrapped' in geo or 'reflect' in geo:
-			nonGeomList.append(i)
-			boundaryList.append(geo)
-		else:
-			geomList.append(i)
-			shapeList.append(geo)
-	return nonGeomList, boundaryList, geomList, shapeList
+    """
+    Get indexes of parameters from paramList
+    based on whether parameter should be sampled in
+    physical (non-geometric) or geometric space,
+    and add these to two lists which can be used
+    to index parameter array
+    """
+    nonGeomList = []
+    geomList = []
+    shapeList = []
+    boundaryList = []
+    for i, geo in enumerate(paramGeomList):
+        if 'vanilla' in geo or 'wrapped' in geo or 'reflect' in geo:
+            nonGeomList.append(i)
+            boundaryList.append(geo)
+        else:
+            geomList.append(i)
+            shapeList.append(geo)
+    return nonGeomList, boundaryList, geomList, shapeList
+
 
 def splitGeomShapes(geomList, shapeList):
-	"""
-	Splits geomList into three separate lists by shape.
-	Assumes that if shape in shapeList is torus or sphere, and an element of
-	shapeList is one of these shapes, then the next parameter in
-	geomList corresponds to second parameter of that pair.
-	Hence for this to work, parameters in geomList corresponding to 3-d shapes
-	should be paired in order
-	"""
-	circleList = []
-	torusList = []
-	sphereList = []
-	i = 0
-	while i < len(geomList):
-		if 'circle' in shapeList[i]:
-			circleList.append(geomList[i])
-		elif 'torus' in shapeList:
-			torusList.append(geomList[i])
-			torusList.append(geomList[i+1])
-			i += 1
-		elif 'sphere' in shapeList:
-			sphereList.append(geomList[i])
-			sphereList.append(geomList[i+1])
-			i += 1
-		i += 1
-	return circleList, torusList, sphereList
+    """
+    Splits geomList into three separate lists by shape.
+    Assumes that if shape in shapeList is torus or sphere, and an element of
+    shapeList is one of these shapes, then the next parameter in
+    geomList corresponds to second parameter of that pair.
+    Hence for this to work, parameters in geomList corresponding to 3-d shapes
+    should be paired in order
+    """
+    circleList = []
+    torusList = []
+    sphereList = []
+    i = 0
+    while i < len(geomList):
+        if 'circle' in shapeList[i]:
+            circleList.append(geomList[i])
+        elif 'torus' in shapeList:
+            torusList.append(geomList[i])
+            torusList.append(geomList[i + 1])
+            i += 1
+        elif 'sphere' in shapeList:
+            sphereList.append(geomList[i])
+            sphereList.append(geomList[i + 1])
+            i += 1
+        i += 1
+    return circleList, torusList, sphereList
+
 
[…]
