+++ gns/recurrence_calculations.py
@@ -1,255 +1,308 @@
-#import standard modules 
+# import standard modules
 import numpy as np
 
-#import custom modules
+# import custom modules
 from . import calculations
 
-###########Updating expected values of Z, X and H functions (recurrence relations of mine, Handley and Skilling)
+# Updating expected values of Z, X and H functions (recurrence relations
+# of mine, Handley and Skilling)
 
-def updateZnXMoments(nLive, EofZ, EofZ2, EofZX, EofX, EofX2, LhoodStarOld, LhoodStar, trapezoidalFlag):
-	"""
-	Wrapper around updateZnXM taking into account whether trapezium rule is used or not
-	"""
-	if trapezoidalFlag:
-		EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight = updateZnXM(nLive, EofZ, EofZ2, EofZX, EofX, EofX2, 0.5 * (LhoodStarOld + LhoodStar))
-	else:
-		EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight = updateZnXM(nLive, EofZ, EofZ2, EofZX, EofX, EofX2, LhoodStar)
-	return EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight
+
+def updateZnXMoments(nLive, EofZ, EofZ2, EofZX, EofX, EofX2,
+                     LhoodStarOld, LhoodStar, trapezoidalFlag):
+    """
+    Wrapper around updateZnXM taking into account whether trapezium rule is used or not
+    """
+    if trapezoidalFlag:
+        EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight = updateZnXM(
+            nLive, EofZ, EofZ2, EofZX, EofX, EofX2, 0.5 * (LhoodStarOld + LhoodStar))
+    else:
+        EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight = updateZnXM(
+            nLive, EofZ, EofZ2, EofZX, EofX, EofX2, LhoodStar)
+    return EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight
+
 
 def updateZnXM(nLive, EofZ, EofZ2, EofZX, EofX, EofX2, L):
-	"""
-	Update moments of Z and X based on their previous values, expected value of random variable t and Lhood value ((L_i + L_i-1) / 2. in case of trapezium rule). 
-	Used to calculate the mean and standard deviation of Z, and thus of log(Z) as well 
-	TODO: CONSIDER KEETON NON-RECURSIVE METHOD
-	"""
-	Eoft, Eoft2, Eof1mt, Eof1mt2 = calculations.calcEofts(nLive)
-	EofZ, EofWeight = updateEofZ(EofZ, Eof1mt, EofX, L)
-	EofZ2 = updateEofZ2(EofZ2, Eof1mt, EofZX, Eof1mt2, EofX2, L)
-	EofZX = updateEofZX(Eoft, EofZX, Eoft2, EofX2, L) 
-	EofX2 = updateEofX2(Eoft2, EofX2)
-	EofX = updateEofX(Eoft, EofX)
-	return EofZ, EofZ2, EofZX, EofX, EofX2, EofWeight
+    """
+    Update moments of Z and X based on their previous values, expected value of random variable t and Lhood value ((L_i + L_i-1) / 2. in case of trapezium rule).
+    Used to calculate the mean and standard deviation of Z, and thus of log(Z) as well
+    TODO: CONSIDER KEETON NON-RECURSIVE METHOD
+    """
+    Eoft, Eoft2, Eof1mt, Eof1mt2 = calculations.calcEofts(nLive)
[…]
