### 2019-11-09 8cb1b19
Ran yapf to make pep8 compliant
