+++ gns/ns_end_funcs.py
@@ -8,8 +8,22 @@ from . import output
 # final contribution to NS sampling functions
 
 
-def getFinalContributionLog(verbose, ZLiveType, trapezoidalFlag, nFinal, logEofZ, logEofZ2, logEofX, logEofWeights,
-                            H, livePointsPhys, livePointsLLhood, avLLhood, liveLLhoodMax, liveMaxIndex, LLhoodStar, errorEval='recursive'):
+def getFinalContributionLog(verbose,
+                            ZLiveType,
+                            trapezoidalFlag,
+                            nFinal,
+                            logEofZ,
+                            logEofZ2,
+                            logEofX,
+                            logEofWeights,
+                            H,
+                            livePointsPhys,
+                            livePointsLLhood,
+                            avLLhood,
+                            liveLLhoodMax,
+                            liveMaxIndex,
+                            LLhoodStar,
+                            errorEval='recursive'):
     """
     Get final contribution from livepoints after NS loop has ended. Way of estimating final contribution is dictated by ZLiveType.
     Also updates H value and gets final weights (and physical values) for posterior
@@ -17,49 +31,68 @@ def getFinalContributionLog(verbose, ZLiveType, trapezoidalFlag, nFinal, logEofZ
     NOTE: for standard quadrature summation, average Lhood and average X give same values of Z (averaging over X is equivalent to averaging over L). However, correct posterior weights are given by latter method, and Z errors are different in both cases
     """
     livePointsLLhood = checkIfAveragedLhood(
-        nFinal, livePointsLLhood, avLLhood)  # only relevant for 'average' ZLiveType
+        nFinal, livePointsLLhood,
+        avLLhood)  # only relevant for 'average' ZLiveType
     if 'average' in ZLiveType:
         LLhoodsFinal = np.concatenate(
             (np.array([LLhoodStar]), livePointsLLhood))
         logEofZOld = logEofZ
         logEofZ2Old = logEofZ2
         for i in range(
-                nFinal):  # add weight of each remaining live point incrementally so H can be calculated easily (according to formulation given in Skilling)
+                nFinal
+        ):  # add weight of each remaining live point incrementally so H can be calculated easily (according to formulation given in Skilling)
             logEofZLive, logEofZ2Live, logEofWeightLive = recurrence_calculations.updateLogZnXMomentsFinal(
-                nFinal, logEofZOld, logEofZ2Old, logEofX, LLhoodsFinal[i], LLhoodsFinal[i + 1], trapezoidalFlag, errorEval)
+                nFinal, logEofZOld, logEofZ2Old, logEofX, LLhoodsFinal[i],
+                LLhoodsFinal[i + 1], trapezoidalFlag, errorEval)
             logEofWeights.append(logEofWeightLive)
-            H = recurrence_calculations.updateHLog(
-                H, logEofWeightLive, logEofZLive, LLhoodsFinal[i + 1], logEofZOld)
+            H = recurrence_calculations.updateHLog(H, logEofWeightLive,
+                                                   logEofZLive,
+                                                   LLhoodsFinal[i + 1],
+                                                   logEofZOld)
             logEofZOld = logEofZLive
             logEofZ2Old = logEofZ2Live
             if verbose:
-                output.printFinalLivePoints(
-                    i, livePointsPhys[i], LLhoodsFinal[i + 1], ZLiveType, 'log')
+                output.printFinalLivePoints(i, livePointsPhys[i],
[…]
