+++ gns/nested_run.py
@@ -18,8 +18,14 @@ from . import geom_sampler
 # nested run functions
 
 
-def NestedRun(priorFunc, invPriorFunc, LhoodFunc, paramNames,
-              targetSupport, setupDict, LLhoodFunc=None, return_vals=False):
+def NestedRun(priorFunc,
+              invPriorFunc,
+              LhoodFunc,
+              paramNames,
+              targetSupport,
+              setupDict,
+              LLhoodFunc=None,
+              return_vals=False):
     """
     Wrapper around linear and log nested run functions.
 
@@ -45,15 +51,20 @@ def NestedRun(priorFunc, invPriorFunc, LhoodFunc, paramNames,
 
     """
     if setupDict['space'] == 'linear':
-        return NestedRunLinear(priorFunc, invPriorFunc, LhoodFunc,
-                               paramNames, targetSupport, setupDict, return_vals)
+        return NestedRunLinear(priorFunc, invPriorFunc, LhoodFunc, paramNames,
+                               targetSupport, setupDict, return_vals)
     elif setupDict['space'] == 'log':
-        return NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc,
-                            paramNames, targetSupport, setupDict, return_vals)
+        return NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames,
+                            targetSupport, setupDict, return_vals)
 
 
-def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames,
-                 targetSupport, setupDict, return_vals=False):
+def NestedRunLog(priorFunc,
+                 invPriorFunc,
+                 LLhoodFunc,
+                 paramNames,
+                 targetSupport,
+                 setupDict,
+                 return_vals=False):
     """
     function which completes a NS run. parameters of priors and likelihood need to be specified, as well as a flag indication type of prior for each dimension and the pdf for the lhood.
     setupDict contains other setup parameters such as termination type & factor, method of finding new livepoint, details of how weights are calculated, how final Z contribution is added,
@@ -125,11 +136,12 @@ def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames,
         # update expected values of moments of X and Z, and get posterior
         # weights
         logEofZNew, logEofZ2, logEofZX, logEofX, logEofX2, logEofWeight = recurrence_calculations.updateLogZnXMoments(
-            nLive, logEofZ, logEofZ2, logEofZX, logEofX, logEofX2, LLhoodStarOld, LLhoodStar, setupDict['trapezoidalFlag'])
+            nLive, logEofZ, logEofZ2, logEofZX, logEofX, logEofX2,
+            LLhoodStarOld, LLhoodStar, setupDict['trapezoidalFlag'])
         logEofXArr.append(logEofX)
         logEofWeights.append(logEofWeight)
-        H = recurrence_calculations.updateHLog(
-            H, logEofWeight, logEofZNew, LLhoodStar, logEofZ)
+        H = recurrence_calculations.updateHLog(H, logEofWeight, logEofZNew,
+                                               LLhoodStar, logEofZ)
         logEofZ = logEofZNew  # update evidence part II
         # WARNING, VIEWING A NUMPY SLICE (IE NOT USING NP.COPY) DOES NOT CREATE A COPY AND SO A-POSTORI CHANGES TO ARRAY WILL AFFECT PREVIOUSLY SLICED ARRAY
         # USE NP.MAY_SHARE_MEMORY(A, B) TO SEE IF ARRAYS SHARE MEMORY, PYTHON
@@ -141,19 +153,29 @@ def NestedRunLog(priorFunc, invPriorFunc, LLhoodFunc, paramNames,
[…]
