+++ gns/main.py
@@ -10,5 +10,6 @@ import input
 def main():
     pass
 
+
 if __name__ == '__main__':
     main()

[…]
