### 2019-11-09 #1809 @SuperKam91
Thank you for you reply @zhampel, that seems very strange. Did you pull from PyPi? I haven't added any of the the review-changes to that version of the package yet, as I figured I would do it at the end of the review process (to avoid having to keep releasing new versions). Please use the version in the github repo for now.
If not I can only think that you looked between me posting the above comment and me uploaded the changes, but they are definitely present in the repo now.

@ziotom78 thank you for the heads up, those are truly great packages! Consider my code now *almost* fully PEP8 compliant- the only places where it isn't is for lines which include comments, which the packages cannot linebreak. I ran the pep8 test and got the following output:
1106    E501 line too long (96 > 79 characters)
Hopefully you will spare me the pain of line breaking my 1106 lines of comments!
