### 2019-11-10 27b755a
Improved docs for toy_models and prob_funcs
