### 2019-11-10 92a9d0f
Improved docs for toy_models and prob_funcs
