### 2019-11-14 #1809 @mbobra
@SuperKam91 There are tools like [autopep8](https://pypi.org/project/autopep8/) that will automatically format your code for you so you don't have to do anything by hand :)
