### 2019-11-19 #1809 @SuperKam91
> @SuperKam91 There are tools like [autopep8](https://pypi.org/project/autopep8/) that will automatically format your code for you so you don't have to do anything by hand :)

Hi @mbobra as I mentioned above, I have already used tools including ‘autopep8’, ‘black’ and ’yapf’ to arrive at the current state. It is fully pep8 compliant apart from my comment lines being over the 80 characters per line limit, and also my doc strings being over the limit. Understandably, the tools cannot correctly break all the commented lines, especially lines which have comments proceeding code.  But since these are just comments I don’t think it should do much harm.
