### 2019-12-09 #1809 @zhampel
@SuperKam91

I apologize for my delay in coming back to your submission.

Thanks for addressing most of the issues. However, I still feel like leaving extremely long comments in **code** sets a negative precedent for JOSS accepted submissions. You wrote the following:
> But since these are just comments I don’t think it should do much harm.

yet, I see this massive single line (327) in tests/test.py:

`setupDict = {'verbose': True, 'trapezoidalFlag': False, 'ZLiveType':'average X', terminationType':'evidence', 'terminationFactor': 0.01, 'sampler': None, 'outputFile': None, 'space':'linear', 'paramGeomList':None}
`

among others. While I agree that long docstrings do not do much harm, I argue that if you want your novel code to be useful to others, with which a publication through JOSS facilitates and which provides authors official credit, then it's best to abide by coding conventions. They are there for a reason, to be helpful for others. And that's why there are automated tools to implement those conventions.

@mbobra
I have reluctantly checked all the boxes and added an issue ticket for the long lines (a more appropriate documentation for this discussion). The state of this codebase is not up to my standards but it is within the guidelines for publication. I would like to stress that reviewers volunteer their time for submissions, and if authors do not take reviewers' time and experience into consideration, then I think it also sets a poor precedent for acceptance. This can affect not just the reviewing process, but more importantly the end-users in the open-source community which we serve.
