### 2019-12-10 #1809 @mbobra
@zhampel Thank you for your time and your well-considered review. I totally and completely understand your comments. We at JOSS are trying to strike some kind of reasonable balance between incentivizing people to write open source software (with a reasonable amount of time and effort and also) and adhering to the most ideal coding standards. I personally agree with you that long comments don't adhere to coding conventions, but I'm also bound (like you) to follow JOSS' guidelines. I hope you feel positive about your experience reviewing for JOSS. And if you feel like the JOSS guidelines should be modified, [please open an issue](https://github.com/openjournals/joss/issues/new)! Thank you sincerely for your time and effort ☀️
