### 2019-12-10 #1809 @mbobra
@SuperKam91 We're almost there! Can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that in your README.md file? After that I think we're done 🎉
