### 2019-12-10 #1809 @mbobra
@SuperKam91 I appreciate your effort; I think pep8 compliance would make the code much more legible. Could you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that in your README.md file? Please make sure the title and author list of the Zenodo entry match the title and author list of the JOSS paper. Thank you!
