+++ docs/index.rst
@@ -17,7 +17,7 @@ Welcome to Geometric nested sampling's documentation!
 Version
 =======
 
-Version: 1.0
+Version: 1.01
 
 Overview
 ========
