+++ README.md
@@ -1,8 +1,8 @@
-![](https://zenodo.org/badge/DOI/10.5281/zenodo.3569956.svg)
-
 # geometric nested sampling
 
-The README below as well as the package API can be found at: [gns read the docs](https://geometric-nested-sampling.readthedocs.io/en/latest/)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3569956.svg)](https://doi.org/10.5281/zenodo.3569956)
+
+The README below as well as the package API can be found at: [gns read the docs](https://geometric-nested-sampling.readthedocs.io/en/latest/).
 
 ## Version
 
