### 2019-12-11 #1809 @mbobra
@SuperKam91 Could you please include DOIs for all your references? So for example, the first entry in paper.bib would look like this:

```
@article{chib1995understanding,
  title={Understanding the metropolis-hastings algorithm},
  author={Chib, Siddhartha and Greenberg, Edward},
  journal={The american statistician},
  volume={49},
  number={4},
  pages={327--335},
  year={1995},
  publisher={Taylor \& Francis Group},
  doi={10.1080/00031305.1995.10476177}
}
```

Thank you 😄
