### 2019-12-12 19a3998
Update paper.bib
