+++ paper.bib
@@ -7,6 +7,7 @@
   pages={327--335},
   year={1995},
   publisher={Taylor \& Francis Group}
+  doi={10.1080/00031305.1995.10476177}
 }
 
 @article{javid2018physical,
@@ -18,6 +19,7 @@
   pages={3529--3544},
   year={2018},
   publisher={Oxford University Press}
+  doi={10.1093/mnras/sty3115}
 }
 
 @article{javid2018comparison,
@@ -25,6 +27,7 @@
   author={Javid, Kamran and Perrott, Yvette C and Hobson, Michael P and Olamaie, Malak and Rumsey, Clare and Saunders, Richard DE},
   journal={arXiv preprint arXiv:1805.01968},
   year={2018}
+  doi={10.17863/CAM.38865}
 }
 
 @article{perrott2019sunyaev,
@@ -36,6 +39,7 @@
   pages={2116--2128},
   year={2019},
   publisher={Oxford University Press}
+  doi={10.1093/mnras/stz826}
 }
 
 @article{javid2019nested,
@@ -43,6 +47,7 @@
   author={Javid, Kamran},
   journal={arXiv preprint arXiv:1905.09110},
   year={2019}
+  doi={}
 }
 
 @phdthesis{javid2019physical,
@@ -50,6 +55,7 @@
   author={Javid, Kamran},
   year={2019},
   school={University of Cambridge}
+  doi={10.17863/CAM.40616}
 }
 
 @article{javid2019physical,
@@ -61,6 +67,7 @@
   pages={3135--3148},
   year={2019},
   publisher={Oxford University Press}
+  doi={10.1093/mnras/stz2341}
 }
 
 @article{feroz2009bayesian,
@@ -72,6 +79,7 @@
   pages={2049--2060},
   year={2009},
   publisher={Blackwell Publishing Ltd Oxford, UK}
+  doi={10.1111/j.1365-2966.2009.15247.x}
 }
 
 @article{feroz2009multinest,
@@ -83,6 +91,7 @@
   pages={1601--1614},
   year={2009},
   publisher={Blackwell Publishing Ltd Oxford, UK}
+  doi={10.1111/j.1365-2966.2009.14548.x}
 }
 
 @article{skilling2006nested,
@@ -94,6 +103,7 @@
   pages={833--859},
   year={2006},
   publisher={International Society for Bayesian Analysis}
+  doi={10.1214/06-BA127}
 }
 
 @article{handley2015polychord,
@@ -105,6 +115,7 @@
   pages={4384--4398},
   year={2015},
   publisher={Oxford University Press}
+  doi={10.1093/mnras/stv1911}
 }
 
 @article{handley2015polychord,
@@ -116,6 +127,7 @@
   pages={L61--L65},
   year={2015},
   publisher={Oxford University Press}
+  doi={10.1093/mnrasl/slv047}
 }
 
 @article{feroz2008multimodal,
@@ -127,6 +139,7 @@
   pages={449--463},
   year={2008},
   publisher={Blackwell Publishing Ltd Oxford, UK}
+  doi={10.1111/j.1365-2966.2007.12353.x}
 }
 @article{hickish2018digital,
   title={A digital correlator upgrade for the Arcminute MicroKelvin Imager},
@@ -137,4 +150,5 @@
   pages={5677--5687},
   year={2018},
   publisher={Oxford University Press}
+  doi={10.1093/mnras/sty074}
 }
