### 2020-01-07 #1809 @SuperKam91
Hi @mbobra apologies for the delay in including the reference DOIs. I am basically waiting on my university to provide me with a DOI for one of the cited papers, but they are taking an age to do so. As soon as they provide it I will update and the references should be good to go!
