### 2020-01-24 #1809 @mbobra
@SuperKam91 Do you want to go ahead and publish the paper without the DOI for that one citation? That's fine with me -- we don't need to hold up your publication just for that singular DOI.
