+++ paper.bib
@@ -47,7 +47,7 @@
   author={Javid, Kamran},
   journal={arXiv preprint arXiv:1905.09110},
   year={2019}
-  doi={}
+  doi={10.5281/zenodo.3627542}
 }
 
 @phdthesis{javid2019physical,
