### 2020-01-25 #1809 @SuperKam91
Hi @mbobra in the end I've lost patience with regard to my university providing me the DOI of the reference, so I just created one for it myself using Zenodo. Hopefully this is sufficient, in which case we can proceed.
