+++ paper.bib
@@ -6,7 +6,7 @@
   number={4},
   pages={327--335},
   year={1995},
-  publisher={Taylor \& Francis Group}
+  publisher={Taylor \& Francis Group},
   doi={10.1080/00031305.1995.10476177}
 }
 
@@ -18,7 +18,7 @@
   number={3},
   pages={3529--3544},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/mnras/sty3115}
 }
 
@@ -26,7 +26,7 @@
   title={Comparison of physical and observational galaxy cluster modelling},
   author={Javid, Kamran and Perrott, Yvette C and Hobson, Michael P and Olamaie, Malak and Rumsey, Clare and Saunders, Richard DE},
   journal={arXiv preprint arXiv:1805.01968},
-  year={2018}
+  year={2018},
   doi={10.17863/CAM.38865}
 }
 
@@ -38,7 +38,7 @@
   number={2},
   pages={2116--2128},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/mnras/stz826}
 }
 
@@ -46,7 +46,7 @@
   title={Nested sampling on non-trivial geometries},
   author={Javid, Kamran},
   journal={arXiv preprint arXiv:1905.09110},
-  year={2019}
+  year={2019},
   doi={10.5281/zenodo.3627542}
 }
 
@@ -54,7 +54,7 @@
   title={Physical modelling of galaxy clusters and Bayesian inference in astrophysics},
   author={Javid, Kamran},
   year={2019},
-  school={University of Cambridge}
+  school={University of Cambridge},
   doi={10.17863/CAM.40616}
 }
 
@@ -66,7 +66,7 @@
   number={3},
   pages={3135--3148},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/mnras/stz2341}
 }
 
@@ -78,7 +78,7 @@
   number={4},
   pages={2049--2060},
   year={2009},
-  publisher={Blackwell Publishing Ltd Oxford, UK}
+  publisher={Blackwell Publishing Ltd Oxford, UK},
   doi={10.1111/j.1365-2966.2009.15247.x}
 }
 
@@ -90,7 +90,7 @@
   number={4},
   pages={1601--1614},
   year={2009},
-  publisher={Blackwell Publishing Ltd Oxford, UK}
+  publisher={Blackwell Publishing Ltd Oxford, UK},
   doi={10.1111/j.1365-2966.2009.14548.x}
 }
 
@@ -102,7 +102,7 @@
   number={4},
   pages={833--859},
   year={2006},
-  publisher={International Society for Bayesian Analysis}
+  publisher={International Society for Bayesian Analysis},
   doi={10.1214/06-BA127}
 }
 
@@ -114,7 +114,7 @@
   number={4},
   pages={4384--4398},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/mnras/stv1911}
 }
 
@@ -126,7 +126,7 @@
   number={1},
   pages={L61--L65},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/mnrasl/slv047}
 }
 
@@ -138,7 +138,7 @@
   number={2},
   pages={449--463},
   year={2008},
-  publisher={Blackwell Publishing Ltd Oxford, UK}
+  publisher={Blackwell Publishing Ltd Oxford, UK},
   doi={10.1111/j.1365-2966.2007.12353.x}
 }
[…]
