### 2020-01-27 #1809 @mbobra
@openjournals/joss-eics the paper is ready for final processing
