### 2020-01-29 #1809 @Kevin-Mattheus-Moerman
@SuperKam91 I am the AEiC helping to process this submission. Below are some remaining minor points to work on in relation to your paper.

- [ ] The first sentence is quite difficult to read. Can you consider rephrasing, splitting into multiple sentences, or adding more comma's?
- [ ] Please change `...present here is a based on...` to `...present here is based on...`
- [ ] Fix typo in `intutive` which should be `intuitive`
- [ ] Fix typo in `equivlance` `equivalance`
- [ ] Consider changing `Figures 6 - 9 of the paper.... ` to `Figures 6 - 9 of that paper.... ` which to me makes it a bit clearer that you are referring to the before mentioned study not the current paper.
- [ ] Please add comma in `To prevent this one must sample a large number...` i.e. `To prevent this, one must sample a large number...`
- [ ] Should `This solution can be problematic if...` be rephrased to `Obtaining this solution can be problematic if...`?
- [ ] Can you replace `...is ‘wrapped’ around...` with `...is "wrapped" around...`, so double quotation symbols are used consistently throughout the text?
- [ ] Please spell out United Kingdom in your affiliations.
- [ ] This paper is about to be processed for acceptance in JOSS. I recommend you also proofread the text thoroughly one last time.
