+++ paper.md
@@ -11,9 +11,9 @@ authors:
     orcid: 0000-0002-2089-3964
     affiliation: "1, 2"
 affiliations:
- - name: Astrophysics Group, Cavendish Laboratory, JJ Thomson Avenue, Cambridge CB3 0HE, UK
+ - name: Kavli Institute for Cosmology Cambridge, Madingley Road, Cambridge, CB3 0HA, United Kingdom
    index: 1
- - name: Kavli Institute for Cosmology Cambridge, Madingley Road, Cambridge, CB3 0HA, UK
+- name: Astrophysics Group, Cavendish Laboratory, JJ Thomson Avenue, Cambridge CB3 0HE, United Kingdom
    index: 2
 date: 2 October 2019
 bibliography: paper.bib
@@ -21,9 +21,9 @@ bibliography: paper.bib
 
 # Abstract
 
-Metropolis Hastings nested sampling evolves a Markov chain from a current livepoint and accepts new points along the chain according to a version of the Metropolis Hastings acceptance ratio modified to satisfy the likelihood constraint, characteristic of nested sampling algorithms.
-The geometric nested sampling algorithm I present here is a based on the Metropolis Hastings method, but treats parameters as though they represent points on certain geometric objects, namely circles, tori and spheres. For parameters which represent points on a circle or torus, the trial distribution is "wrapped" around the domain of the posterior distribution such that samples cannot be rejected automatically when evaluating the Metropolis ratio due to being outside the sampling domain. Furthermore, this enhances the mobility of the sampler. 
-For parameters which represent coordinates on the surface of a sphere, the algorithm transforms the parameters into a Cartesian coordinate system before sampling which again makes sure no samples are automatically rejected, and provides a physically intutive way of the sampling the parameter space.
+Metropolis Hastings nested sampling evolves a Markov chain, accepting new points along the chain according to a version of the Metropolis Hastings acceptance ratio, which has been modified to satisfy the nested sampling likelihood constraint.
+The geometric nested sampling algorithm I present here is based on the Metropolis Hastings method, but treats parameters as though they represent points on certain geometric objects, namely circles, tori and spheres. For parameters which represent points on a circle or torus, the trial distribution is "wrapped" around the domain of the posterior distribution such that samples cannot be rejected automatically when evaluating the Metropolis ratio due to being outside the sampling domain. Furthermore, this enhances the mobility of the sampler. 
+For parameters which represent coordinates on the surface of a sphere, the algorithm transforms the parameters into a Cartesian coordinate system before sampling which again makes sure no samples are automatically rejected, and provides a physically intuitive way of the sampling the parameter space.
 
 # Bayesian inference and nested sampling
 
@@ -45,26 +45,26 @@ $$
 \mathcal{Z} = \int_{0}^{1} \mathcal{L}(X) \mathrm{d}X.
 $$
 
-The integral extends over the region(s) of the parameter space contained within the iso-likelihood contour $\mathcal{L}\left(\mathbf{\theta}\right) = \mathcal{L}$ (see @javid2019physical for a proof of the equivlance of both evidence integrals).
+The integral extends over the region(s) of the parameter space contained within the iso-likelihood contour $\mathcal{L}\left(\mathbf{\theta}\right) = \mathcal{L}$ (see @javid2019physical for a proof of the equivalence of both evidence integrals).
 
[…]
