### 2020-01-30 #1809 @SuperKam91
Hi @Kevin-Mattheus-Moerman, I can confirm I have addressed the changes proposed above. Furthermore I have given the paper a final proofread.
