### 2020-01-30 #1809 @Kevin-Mattheus-Moerman
@SuperKam91 did you have the ` symbols in the title field for the YAML content at the top of your paper there before?
```yaml
title: 'Geometric nested sampling: sampling from distributions defined on non-trivial geometries'
```
Perhaps change it to the following and try again:
```yaml
title: Geometric nested sampling: sampling from distributions defined on non-trivial geometries
```
