### 2020-01-30 #1809 @arfon
@SuperKam91 - this PR will fix your paper compilation: https://github.com/SuperKam91/gns/pull/6
