### 2020-02-04 #1809 @SuperKam91
Many thanks all for your help, I have made a small donation to JOSS in the hope that you all can keep up the good work in the future.

I have noticed a small typo in the references of the paper. Is it possible to update the publication now I have ratified it?

If not absolutely no worries, I will just include it in the arXiv version instead.
