### 2020-02-04 #1809 @arfon
> Many thanks all for your help, I have made a small donation to JOSS in the hope that you all can keep up the good work in the future.

❤️ Many thanks!

> I have noticed a small typo in the references of the paper. Is it possible to update the publication now I have ratified it?

Yes, please go ahead and make the fix in the `master` branch of your repository and mention me here (@arfon) when you've done this.
