### 2020-02-04 #1809 @SuperKam91
> > Many thanks all for your help, I have made a small donation to JOSS in the hope that you all can keep up the good work in the future.
>
>  Many thanks!
>
> > I have noticed a small typo in the references of the paper. Is it possible to update the publication now I have ratified it?
>
> Yes, please go ahead and make the fix in the `master` branch of your repository and mention me here (@arfon) when you've done this.

Hi @arfon that's great, thank you so much. I have made the changes to `master` branch so we should be good to go. However, when "checking references" above it seems to have cached an old commit (perhaps I didn't wait long enough but I don't think this is usually an issue) which doesn't include the fix, so let me know if you have any problems with this.
