### 2020-02-05 #1809 @arfon
OK, this should be fixed now. The PDF might take a few hours to show up as fixed online due to caching.

Also, I think you're right, there's some kind of caching issue with Whedon too that I need to look into.
